+++ test/MoqxRelayFetchTests.cpp
@@ -1,0 +1,68 @@
+/*
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ * Originally from github.com/facebookexperimental/moxygen.
+ * See deps/moxygen/LICENSE for the original license terms.
+ *
+ * Copyright (c) OpenMOQ contributors.
+ */
+
+#include "MoqxRelayTestFixture.h"
+
+namespace moxygen::test {
+
+// Test: fetch fallback to subscriptions_ after publisher termination must not
+// crash. When findPublishNamespaceSession returns null (no publishNamespace),
+// fetch falls back to subscriptions_. After onPublishDone, upstream is null
+// but the subscription entry remains if the forwarder has subscribers.
+TEST_F(MoQRelayTest, FetchAfterPublisherTermination) {
+  auto publisherSession = createMockSession();
+  auto subSession = createMockSession();
+  auto fetchSession = createMockSession();
+
+  // Publish WITHOUT publishNamespace so findPublishNamespaceSession returns null
+  // and fetch falls back to subscriptions_
+  auto publishConsumer = doPublish(publisherSession, kTestTrackName, /*addToState=*/false);
+
+  // Subscriber with open subgroup so subscription survives publisher drain
+  auto consumer = createMockConsumer();
+  auto sg = createMockSubgroupConsumer();
+  EXPECT_CALL(*consumer, beginSubgroup(0, 0, _, _))
+      .WillOnce([&sg](uint64_t, uint64_t, uint8_t, bool) {
+        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
+      });
+  auto handle = subscribeToTrack(subSession, kTestTrackName, consumer, RequestID(0));
+  ASSERT_NE(handle, nullptr);
+
+  // Begin subgroup to keep subscriber alive
+  auto subgroupRes = publishConsumer->beginSubgroup(0, 0, 0);
+  ASSERT_TRUE(subgroupRes.hasValue());
+
+  // Publisher terminates — clears upstream but subscription stays
+  EXPECT_CALL(*consumer, publishDone(_))
+      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
+  publishConsumer->publishDone(
+      {RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher ended"}
+  );
+
+  // Fetch from a different session — falls back to subscriptions_, gets null
+  // upstream, then crashes at line 1011 dereferencing null upstreamSession
+  Fetch fetch(RequestID(0), kTestTrackName, AbsoluteLocation{0, 0}, AbsoluteLocation{1, 0});
+  auto fetchConsumer = std::make_shared<NiceMock<MockFetchConsumer>>();
+  withSessionContext(fetchSession, [&]() {
+    auto task = relay_->fetch(std::move(fetch), fetchConsumer);
+    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
+    // Should return an error, not crash
+    EXPECT_FALSE(res.hasValue());
+    EXPECT_EQ(res.error().errorCode, FetchErrorCode::TRACK_NOT_EXIST);
+  });
+
+  // Clean up
+  EXPECT_CALL(*sg, reset(_)).Times(1);
+  subgroupRes.value()->reset(ResetStreamErrorCode::CANCELLED);
+  handle->unsubscribe();
+  removeSession(publisherSession);
+  removeSession(subSession);
+  removeSession(fetchSession);
+}
+
+} // namespace moxygen::test
