+++ test/MoqxRelaySubNsTests.cpp
@@ -1,0 +1,222 @@
+/*
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ * Originally from github.com/facebookexperimental/moxygen.
+ * See deps/moxygen/LICENSE for the original license terms.
+ *
+ * Copyright (c) OpenMOQ contributors.
+ */
+
+#include "MoqxRelayTestFixture.h"
+
+namespace moxygen::test {
+
+TEST_F(MoQRelayTest, SubscribeNamespaceDoesntAddDrainingPublish) {
+  auto publisherSession = createMockSession();
+  auto subscriber1 = createMockSession();
+  auto subscriber2 = createMockSession();
+
+  // Subscriber 1 subscribes to publishNamespaces
+  auto handle1 = doSubscribeNamespace(subscriber1, kTestNamespace, /*addToState=*/false);
+
+  // Publish first track - subscriber 1 should receive it
+  auto mockConsumer1 = createMockConsumer();
+  EXPECT_CALL(*subscriber1, publish(testing::_, testing::_))
+      .WillOnce([mockConsumer1](const auto& /*pubReq*/, auto /*subHandle*/) {
+        return Subscriber::PublishResult(Subscriber::PublishConsumerAndReplyTask{
+            mockConsumer1,
+            []() -> folly::coro::Task<folly::Expected<PublishOk, PublishError>> {
+              co_return PublishOk{
+                  /*requestID=*/RequestID(1),
+                  /*forward=*/true,
+                  /*subscriberPriority=*/0,
+                  /*groupOrder=*/GroupOrder::OldestFirst,
+                  /*locType=*/LocationType::LargestObject,
+                  /*start=*/std::nullopt,
+                  /*endGroup=*/std::nullopt
+              };
+            }()
+        });
+      });
+
+  EXPECT_CALL(*mockConsumer1, beginSubgroup(_, _, _, _))
+      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
+        auto sg = std::make_shared<NiceMock<MockSubgroupConsumer>>();
+        EXPECT_CALL(*sg, endOfSubgroup()).WillOnce(testing::Return(folly::unit));
+        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
+      });
+
+  // Begin a subgroup for ongoing publish activity
+  auto pubConsumer = doPublish(
+      publisherSession,
+      FullTrackName{kTestNamespace, "track_stream"},
+      /*addToState=*/false
+  );
+  // TODO: bug subscriber not added until next loop?
+  exec_->drive();
+  auto subgroupRes = pubConsumer->beginSubgroup(0, 0, 0);
+  EXPECT_TRUE(subgroupRes.hasValue());
+  auto subgroup = *subgroupRes;
+
+  // publisher ends subscription
+  EXPECT_CALL(*mockConsumer1, publishDone(testing::_));
+  EXPECT_TRUE(
+      pubConsumer->publishDone({RequestID(1), PublishDoneStatusCode::TRACK_ENDED, 0, "track ended"})
+          .hasValue()
+  );
+  subgroup->endOfSubgroup();
+
+  // Subscriber 2 subscribes to publishNamespaces but doesn't get finished track
+  doSubscribeNamespace(subscriber2, kTestNamespace);
+
+  // First publish (existing context handles initial publish), now publish a
+  // second track
+  // Expect publish calls on both subscribers, just fail them.
+  EXPECT_CALL(*subscriber1, publish(testing::_, testing::_))
+      .WillOnce([](const auto& /*pubReq*/, auto /*subHandle*/) {
+        return folly::makeUnexpected(PublishError{});
+      });
+
+  EXPECT_CALL(*subscriber2, publish(testing::_, testing::_))
+      .WillOnce([](const auto& /*pubReq*/, auto /*subHandle*/) {
+        return folly::makeUnexpected(PublishError{});
+      });
+
+  auto pubConsumer2 = doPublish(publisherSession, FullTrackName{kTestNamespace, "track_stream_2"});
+  exec_->drive();
+
+  removeSession(publisherSession);
+  removeSession(subscriber1);
+  removeSession(subscriber2);
+}
+
+TEST_F(MoQRelayTest, SubscribeNamespaceEmptyPrefixRejectedPreV16) {
+  // Default session uses kVersionDraftCurrent (draft-14, which is < 16)
+  auto session = createMockSession();
+
+  TrackNamespace emptyNs{{}};
+  SubscribeNamespace subNs;
+  subNs.trackNamespacePrefix = emptyNs;
+
+  withSessionContext(session, [&]() {
+    auto task = relay_->subscribeNamespace(std::move(subNs), nullptr);
+    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
+    ASSERT_FALSE(res.hasValue()
+    ) << "Empty namespace prefix should be rejected for pre-v16 sessions";
+    EXPECT_EQ(res.error().errorCode, SubscribeNamespaceErrorCode::NAMESPACE_PREFIX_UNKNOWN);
+    EXPECT_EQ(res.error().reasonPhrase, "empty");
+  });
+
+  removeSession(session);
+}
+
+TEST_F(MoQRelayTest, SubscribeNamespaceEmptyPrefixAllowedV16) {
+  auto session = createMockSession();
+  // Override the negotiated version to draft-16
+  ON_CALL(*session, getNegotiatedVersion())
+      .WillByDefault(Return(std::optional<uint64_t>(kVersionDraft16)));
+
+  TrackNamespace emptyNs{{}};
+  doSubscribeNamespace(session, emptyNs);
+
+  removeSession(session);
+}
+
+TEST_F(MoQRelayTest, ExactNamespaceSubscriberReceivesPublishNamespace) {
+  auto subscriber = createMockSession();
+  auto publisher = createMockSession();
+
+  // Subscriber subscribes to exact namespace {"test", "namespace"}
+  doSubscribeNamespace(subscriber, kTestNamespace);
+
+  // Expect the subscriber to receive a publishNamespace forwarding when
+  // the publisher announces the same exact namespace
+  EXPECT_CALL(*subscriber, publishNamespace(_, _))
+      .WillOnce(
+          [](PublishNamespace ann, auto) -> folly::coro::Task<Subscriber::PublishNamespaceResult> {
+            EXPECT_EQ(ann.trackNamespace, kTestNamespace);
+            co_return folly::makeUnexpected(PublishNamespaceError{
+                ann.requestID,
+                PublishNamespaceErrorCode::UNINTERESTED,
+                "test"
+            });
+          }
+      );
+
+  // Publisher announces the same exact namespace
+  doPublishNamespace(publisher, kTestNamespace);
+
+  // Drive the executor so the async publishNamespace forwarding runs
+  exec_->drive();
+
+  removeSession(publisher);
+  removeSession(subscriber);
+}
+
+// Bug: when a subscriber with forward=true joins a namespace whose track
+// forwarder is empty, the relay fires REQUEST_UPDATE twice — once explicitly
+// at the if(forwarder->empty()) site and once via forwardChanged() when
+// addSubscriber() increments numForwardingSubscribers from 0 to 1.
+TEST_F(MoQRelayTest, SubscribeNs_ForwardTrue_EmptyForwarder_SingleRequestUpdate) {
+  auto pubSession = createMockSession();
+  doPublishNamespace(pubSession, kTestNamespace);
+  auto mockHandle = makePublishHandle();
+  doPublishWithHandle(pubSession, kTestTrackName, mockHandle);
+
+  // Expect exactly one REQUEST_UPDATE(forward=true).
+  // Before the fix this fires twice.
+  EXPECT_CALL(*mockHandle, requestUpdateCalled(_)).Times(1).WillOnce([](const RequestUpdate& u) {
+    ASSERT_TRUE(u.forward.has_value());
+    EXPECT_TRUE(*u.forward);
+  });
+
+  auto subSession = createMockSession();
+  setupPublishSucceeds(subSession);
+  doSubscribeNamespaceWithForward(subSession, kTestNamespace, /*forward=*/true);
+
+  for (int i = 0; i < 5; i++) {
+    exec_->drive();
+  }
+
+  // Verify before cleanup — cleanup itself legitimately sends forward=false
+  // when the subscriber leaves and the forwarder drains.
+  ASSERT_TRUE(testing::Mock::VerifyAndClearExpectations(mockHandle.get()));
+
+  removeSession(subSession);
+  removeSession(pubSession);
+  for (int i = 0; i < 3; i++) {
+    exec_->drive();
+  }
+}
+
+// Bug: when a subscriber with forward=false joins a namespace whose track
+// forwarder is empty, the relay fires a spurious REQUEST_UPDATE(forward=false)
+// at the if(forwarder->empty()) site — even though the upstream is already at
+// forward=false (set by publish() which found no subscribers).
+TEST_F(MoQRelayTest, SubscribeNs_ForwardFalse_EmptyForwarder_NoRequestUpdate) {
+  auto pubSession = createMockSession();
+  doPublishNamespace(pubSession, kTestNamespace);
+  auto mockHandle = makePublishHandle();
+  doPublishWithHandle(pubSession, kTestTrackName, mockHandle);
+
+  // Expect no REQUEST_UPDATE at all.
+  // Before the fix this fires once with forward=false.
+  EXPECT_CALL(*mockHandle, requestUpdateCalled(_)).Times(0);
+
+  auto subSession = createMockSession();
+  setupPublishSucceeds(subSession);
+  doSubscribeNamespaceWithForward(subSession, kTestNamespace, /*forward=*/false);
+
+  for (int i = 0; i < 5; i++) {
+    exec_->drive();
+  }
+
+  ASSERT_TRUE(testing::Mock::VerifyAndClearExpectations(mockHandle.get()));
+
+  removeSession(subSession);
+  removeSession(pubSession);
+  for (int i = 0; i < 3; i++) {
+    exec_->drive();
+  }
+}
+
+} // namespace moxygen::test
