--- test/MoqxRelayTest.cpp
@@ -1,2937 +1,0 @@
-/*
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- * Originally from github.com/facebookexperimental/moxygen.
- * See deps/moxygen/LICENSE for the original license terms.
- *
- * Copyright (c) OpenMOQ contributors.
- */
-
-#include "MoqxRelayTestFixture.h"
-#include "UpstreamProvider.h"
-
-namespace moxygen::test {
-
-// Test: Basic relay construction
-TEST_F(MoQRelayTest, Construction) {
-  EXPECT_NE(relay_, nullptr);
-}
-
-// Test: Verify allowed namespace prefix is set correctly
-TEST_F(MoQRelayTest, AllowedNamespacePrefix) {
-  // This just verifies the relay can be constructed with a namespace prefix
-  // More detailed testing requires full session setup
-  auto relay2 = std::make_shared<MoqxRelay>(config::CacheConfig{
-      .maxCachedTracks = moxygen::kDefaultMaxCachedTracks,
-      .maxCachedGroupsPerTrack = moxygen::kDefaultMaxCachedGroupsPerTrack,
-  });
-  relay2->setAllowedNamespacePrefix(kTestNamespace);
-  EXPECT_NE(relay2, nullptr);
-}
-
-// Test: MockMoQSession can be created
-TEST_F(MoQRelayTest, MockSessionCreation) {
-  auto mockSession = createMockSession();
-  EXPECT_NE(mockSession, nullptr);
-  EXPECT_NE(mockSession->getExecutor(), nullptr);
-}
-
-// Test: Publish a track through the relay
-TEST_F(MoQRelayTest, PublishSuccess) {
-  auto publisherSession = createMockSession();
-
-  // Publish the namespace
-  doPublishNamespace(publisherSession, kTestNamespace);
-
-  // Publish the track
-  doPublish(publisherSession, kTestTrackName);
-
-  // Cleanup: remove the session from relay to avoid mock leak warning
-  removeSession(publisherSession);
-}
-
-// Namespace tree tests (Prune*, MixedContent*, ActiveChildCount, PublishKeepsNode)
-// are in NamespaceTreeTest.cpp.
-
-// Test: Verify that subgroups are only created at appropriate times
-// Sequence: publish, sub1 joins, beginSubgroup (->1), beginObject (->1),
-// sub2 joins, beginObject (->1,2), sub3 joins, objectPayload (->1,2),
-// endOfSubgroup (->1,2)
-TEST_F(MoQRelayTest, ForwarderOnlyCreatesSubgroupsBeforeObjectData) {
-  auto publisherSession = createMockSession();
-  auto subscriber1 = createMockSession();
-  auto subscriber2 = createMockSession();
-  auto subscriber3 = createMockSession();
-
-  // Set up mock consumers with expectations
-  auto mockConsumer1 = createMockConsumer();
-  auto mockConsumer2 = createMockConsumer();
-  auto mockConsumer3 = createMockConsumer();
-
-  auto setupSubgroupConsumer = [](auto& sg) {
-    ON_CALL(*sg, beginObject(_, _, _, _))
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-    ON_CALL(*sg, objectPayload(_, _))
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(ObjectPublishStatus::IN_PROGRESS)
-        ));
-    ON_CALL(*sg, endOfSubgroup())
-        .WillByDefault(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-  };
-
-  // Subscriber 1 and 2 should get beginSubgroup
-  EXPECT_CALL(*mockConsumer1, beginSubgroup(_, _, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        auto sg = std::make_shared<NiceMock<MockSubgroupConsumer>>();
-        setupSubgroupConsumer(sg);
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  EXPECT_CALL(*mockConsumer2, beginSubgroup(_, _, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        auto sg = std::make_shared<NiceMock<MockSubgroupConsumer>>();
-        setupSubgroupConsumer(sg);
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  // Subscriber 3 joins mid-object and should NOT get beginSubgroup
-  EXPECT_CALL(*mockConsumer3, beginSubgroup(_, _, _, _)).Times(0);
-
-  // Setup: publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Subscriber 1 joins before any data
-  subscribeToTrack(subscriber1, kTestTrackName, mockConsumer1, RequestID(1));
-
-  // Publisher sends data with subscribers joining at different points
-  auto sgRes = publishConsumer->beginSubgroup(0, 0, 0);
-  EXPECT_TRUE(sgRes.hasValue());
-  auto subgroup = *sgRes;
-
-  // First object -> goes to subscriber 1
-  EXPECT_TRUE(subgroup->beginObject(0, 4, 0).hasValue());
-  EXPECT_TRUE(subgroup->objectPayload(folly::IOBuf::copyBuffer("test"), false).hasValue());
-
-  // Subscriber 2 joins after first object
-  subscribeToTrack(subscriber2, kTestTrackName, mockConsumer2, RequestID(2));
-
-  // Second object -> goes to subscribers 1, 2
-  EXPECT_TRUE(subgroup->beginObject(1, 9, 0).hasValue());
-
-  // Subscriber 3 joins mid-object
-  subscribeToTrack(subscriber3, kTestTrackName, mockConsumer3, RequestID(3));
-
-  // Payload and endOfSubgroup -> only to subscribers 1, 2 (NOT 3)
-  EXPECT_TRUE(subgroup->objectPayload(folly::IOBuf::copyBuffer("more data"), false).hasValue());
-  EXPECT_TRUE(subgroup->endOfSubgroup().hasValue());
-
-  removeSession(publisherSession);
-  removeSession(subscriber1);
-  removeSession(subscriber2);
-  removeSession(subscriber3);
-}
-
-// Test: Graceful session draining - comprehensive test of draining behavior
-// This test verifies that when a publisher calls publishDone, subscribers
-// are drained (receive publishDone) but their open subgroups are NOT reset.
-// Draining subscribers should not receive new subgroups and are removed when
-// their last subgroup closes.
-TEST_F(MoQRelayTest, GracefulSessionDraining) {
-  auto publisherSession = createMockSession();
-  std::array<std::shared_ptr<MoQSession>, 3> subscribers;
-  for (auto& s : subscribers) {
-    s = createMockSession();
-  }
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Create subscribers with different subgroup states
-  std::array<std::shared_ptr<MockSubgroupConsumer>, 2> sub0_sgs;
-  std::shared_ptr<MockSubgroupConsumer> sub1_sg0;
-
-  std::array<std::shared_ptr<MockTrackConsumer>, 3> consumers;
-  for (auto& c : consumers) {
-    c = createMockConsumer();
-  }
-
-  // Subscriber1: will have 2 open subgroups (0, 1)
-  for (uint64_t i = 0; i < 2; ++i) {
-    EXPECT_CALL(*consumers[0], beginSubgroup(i, 0, _, _))
-        .WillOnce([this, i, &sub0_sgs](uint64_t, uint64_t, uint8_t, bool) {
-          sub0_sgs[i] = createMockSubgroupConsumer();
-          return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sub0_sgs[i]
-          );
-        });
-  }
-
-  // Subscriber2: will have 1 open subgroup (1)
-  EXPECT_CALL(*consumers[1], beginSubgroup(1, 0, _, _))
-      .WillOnce([this, &sub1_sg0](uint64_t, uint64_t, uint8_t, bool) {
-        sub1_sg0 = createMockSubgroupConsumer();
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sub1_sg0);
-      });
-
-  std::array<std::shared_ptr<SubgroupConsumer>, 2> sgs;
-  subscribeToTrack(subscribers[0], kTestTrackName, consumers[0], RequestID(1));
-  sgs[0] = publishConsumer->beginSubgroup(0, 0, 0).value();
-  subscribeToTrack(subscribers[1], kTestTrackName, consumers[1], RequestID(2));
-  sgs[1] = publishConsumer->beginSubgroup(1, 0, 0).value();
-  subscribeToTrack(subscribers[2], kTestTrackName, consumers[2], RequestID(3));
-
-  // Trigger draining by calling publishDone from publisher
-  // All subscribers should receive publishDone but subgroups should NOT reset
-  // Note: publishDone may be called multiple times (during drain and cleanup)
-  for (auto& consumer : consumers) {
-    EXPECT_CALL(*consumer, publishDone(_))
-        .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-  }
-
-  // Verify subgroups are NOT reset (no reset() calls expected)
-  EXPECT_CALL(*sub0_sgs[0], reset(_)).Times(0);
-  EXPECT_CALL(*sub0_sgs[1], reset(_)).Times(0);
-  EXPECT_CALL(*sub1_sg0, reset(_)).Times(0);
-
-  publishConsumer->publishDone(
-      PublishDone{RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher ended"}
-  );
-
-  // At this point:
-  // - subscriber3 removed (had no subgroups)
-  // - subscriber1 still exists (has 2 open subgroups)
-  // - subscriber2 still exists (has 1 open subgroup)
-
-  // Close subgroup 0 -> subscriber2 removed (was its last subgroup)
-  // subscriber1 still has subgroup 1
-  EXPECT_TRUE(sgs[0]->endOfSubgroup().hasValue());
-
-  // Close subgroup 1 -> subscriber1 removed (was its last subgroup)
-  EXPECT_TRUE(sgs[1]->endOfSubgroup().hasValue());
-
-  removeSession(publisherSession);
-}
-
-// Test: removeSession (via unsubscribe) immediately resets all open subgroups
-TEST_F(MoQRelayTest, RemoveSessionResetsOpenSubgroups) {
-  auto publisherSession = createMockSession();
-  auto subscriber = createMockSession();
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Create subscriber with 2 open subgroups
-  std::array<std::shared_ptr<MockSubgroupConsumer>, 2> sgs;
-  auto consumer = createMockConsumer();
-
-  for (uint64_t i = 0; i < sgs.size(); ++i) {
-    EXPECT_CALL(*consumer, beginSubgroup(i, 0, _, _))
-        .WillOnce([this, i, &sgs](uint64_t, uint64_t, uint8_t, bool) {
-          sgs[i] = createMockSubgroupConsumer();
-          EXPECT_CALL(*sgs[i], object(0, testing::_, testing::_, false))
-              .WillOnce(testing::Return(folly::unit));
-          return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sgs[i]);
-        });
-  }
-
-  auto subHandle =
-      subscribeToTrack(subscriber, kTestTrackName, consumer, RequestID(1), /*addToState=*/false);
-
-  // Publish data to create subgroups and first objects
-  std::array<std::shared_ptr<SubgroupConsumer>, 2> pubSgs;
-  for (uint64_t i = 0; i < pubSgs.size(); ++i) {
-    auto sg = publishConsumer->beginSubgroup(i, 0, 0);
-    ASSERT_TRUE(sg.hasValue());
-    EXPECT_TRUE((*sg)->object(0, test::makeBuf(10)).hasValue());
-    pubSgs[i] = *sg;
-  }
-
-  // Expect both subgroups to be reset when subscriber unsubscribes
-  for (auto& sg : sgs) {
-    EXPECT_CALL(*sg, reset(ResetStreamErrorCode::CANCELLED));
-  }
-
-  // Unsubscribe triggers removeSession which resets all subgroups
-  withSessionContext(subscriber, [&]() { subHandle->unsubscribe(); });
-
-  for (uint64_t i = 0; i < pubSgs.size(); ++i) {
-    // Ensure each published subgroup is properly closed after unsubscribe
-    EXPECT_TRUE(pubSgs[i]->endOfSubgroup().hasValue());
-  }
-  removeSession(subscriber);
-  removeSession(publisherSession);
-}
-
-// Test: Draining subscriber removed when subgroup closes with error
-TEST_F(MoQRelayTest, DrainingSubscriberRemovedOnSubgroupError) {
-  auto publisherSession = createMockSession();
-  auto subscriber = createMockSession();
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Create subscriber with 3 open subgroups (0, 1, 2)
-  std::array<std::shared_ptr<MockSubgroupConsumer>, 3> sgs;
-  auto consumer = createMockConsumer();
-
-  for (uint64_t i = 0; i < sgs.size(); ++i) {
-    EXPECT_CALL(*consumer, beginSubgroup(i, 0, _, _))
-        .WillOnce([this, i, &sgs](uint64_t, uint64_t, uint8_t, bool) {
-          sgs[i] = createMockSubgroupConsumer();
-          return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sgs[i]);
-        });
-  }
-
-  subscribeToTrack(subscriber, kTestTrackName, consumer, RequestID(1));
-
-  // Publish data to create all 3 subgroups
-  std::array<std::shared_ptr<SubgroupConsumer>, 3> pubSgs;
-  for (uint64_t i = 0; i < 3; ++i) {
-    pubSgs[i] = publishConsumer->beginSubgroup(i, 0, 0).value();
-    EXPECT_TRUE((pubSgs[i])->beginObject(0, 10, 0).hasValue());
-  }
-
-  // Drain the subscriber - should mark it draining but not remove
-  // Note: publishDone may be called multiple times (during drain and cleanup)
-  EXPECT_CALL(*consumer, publishDone(_))
-      .Times(AtLeast(1))
-      .WillRepeatedly(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-
-  // Subgroups should NOT be reset during drain
-  EXPECT_CALL(*sgs[0], reset(_)).Times(0);
-  EXPECT_CALL(*sgs[1], reset(_)).Times(0);
-  EXPECT_CALL(*sgs[2], reset(_)).Times(0);
-
-  publishConsumer->publishDone(
-      PublishDone{RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher ended"}
-  );
-
-  // Now close subgroups one by one
-  // Test 1: Send error on subgroup 0 -> subscriber NOT removed (still has 1, 2)
-  EXPECT_CALL(*sgs[0], reset(ResetStreamErrorCode::INTERNAL_ERROR));
-  pubSgs[0]->reset(ResetStreamErrorCode::INTERNAL_ERROR);
-
-  // Test 2: Close subgroup 1 normally -> subscriber NOT removed (still has 2)
-  EXPECT_CALL(*sgs[1], objectPayload(testing::_, true));
-  EXPECT_TRUE(pubSgs[1]->objectPayload(test::makeBuf(10), true));
-
-  // Test 3: Send error on subgroup 2 -> subscriber IS removed (was last)
-  EXPECT_CALL(*sgs[2], reset(ResetStreamErrorCode::INTERNAL_ERROR));
-  pubSgs[2]->reset(ResetStreamErrorCode::INTERNAL_ERROR);
-
-  removeSession(publisherSession);
-}
-
-// Test: Subscriber that ends subscription doesn't receive subsequent objects
-// Sequence: publish, sub1 subscribes, beginSubgroup, sub1 ends (publishDone),
-// sub2 subscribes, send object -> only goes to sub1
-TEST_F(MoQRelayTest, SubscriberUnsubscribeDoesNotReceiveNewObjects) {
-  auto publisherSession = createMockSession();
-  auto subscriber1 = createMockSession();
-  auto subscriber2 = createMockSession();
-
-  // Set up mock consumers
-  auto mockConsumer1 = createMockConsumer();
-  auto mockConsumer2 = createMockConsumer();
-
-  // Sub1 should get beginSubgroup
-  EXPECT_CALL(*mockConsumer1, beginSubgroup(_, _, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        auto sg = std::make_shared<NiceMock<MockSubgroupConsumer>>();
-        EXPECT_CALL(*sg, beginObject(_, _, _, _)).WillOnce(Return(folly::unit));
-        EXPECT_CALL(*sg, reset(_));
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  // Sub2 should also get beginSubgroup
-  EXPECT_CALL(*mockConsumer2, beginSubgroup(_, _, _, _)).Times(0);
-
-  // Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName, /*addToState=*/false);
-
-  // Subscriber 1 joins
-  subscribeToTrack(subscriber1, kTestTrackName, mockConsumer1, RequestID(1));
-
-  // Begin subgroup - sub1 receives it
-  auto sgRes = publishConsumer->beginSubgroup(0, 0, 0);
-  EXPECT_TRUE(sgRes.hasValue());
-  auto subgroup = *sgRes;
-
-  // publisher ends subscription
-  EXPECT_CALL(*mockConsumer1, publishDone(testing::_));
-  EXPECT_TRUE(publishConsumer
-                  ->publishDone({RequestID(1), PublishDoneStatusCode::TRACK_ENDED, 0, "track ended"}
-                  )
-                  .hasValue());
-
-  // Subscriber 2 joins after publishDone
-  subscribeToTrack(
-      subscriber2,
-      kTestTrackName,
-      mockConsumer2,
-      RequestID(2),
-      /*addToState=*/false,
-      SubscribeErrorCode::INTERNAL_ERROR
-  );
-
-  // Send object - should only go to subscriber 1 (sub2 is gone)
-  EXPECT_TRUE(subgroup->beginObject(0, 4, 0).hasValue());
-  subgroup->reset(ResetStreamErrorCode::SESSION_CLOSED);
-
-  removeSession(publisherSession);
-  removeSession(subscriber1);
-  removeSession(subscriber2);
-}
-
-// Test: SubscribeNamespace only receives publishes while active
-// Sequence: subscribeNamespace (sub1), publish (sub1 gets it),
-// beginSubgroup, publishDone, subscribeNamespace (sub2), new publish
-// (only sub2 gets it)
-TEST_F(MoQRelayTest, SubscribeNamespaceDoesntAddDrainingPublish) {
-  auto publisherSession = createMockSession();
-  auto subscriber1 = createMockSession();
-  auto subscriber2 = createMockSession();
-
-  // Subscriber 1 subscribes to publishNamespaces
-  auto handle1 = doSubscribeNamespace(subscriber1, kTestNamespace, /*addToState=*/false);
-
-  // Publish first track - subscriber 1 should receive it
-  auto mockConsumer1 = createMockConsumer();
-  EXPECT_CALL(*subscriber1, publish(testing::_, testing::_))
-      .WillOnce([mockConsumer1](const auto& /*pubReq*/, auto /*subHandle*/) {
-        return Subscriber::PublishResult(Subscriber::PublishConsumerAndReplyTask{
-            mockConsumer1,
-            []() -> folly::coro::Task<folly::Expected<PublishOk, PublishError>> {
-              co_return PublishOk{
-                  /*requestID=*/RequestID(1),
-                  /*forward=*/true,
-                  /*subscriberPriority=*/0,
-                  /*groupOrder=*/GroupOrder::OldestFirst,
-                  /*locType=*/LocationType::LargestObject,
-                  /*start=*/std::nullopt,
-                  /*endGroup=*/std::nullopt
-              };
-            }()
-        });
-      });
-
-  EXPECT_CALL(*mockConsumer1, beginSubgroup(_, _, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        auto sg = std::make_shared<NiceMock<MockSubgroupConsumer>>();
-        EXPECT_CALL(*sg, endOfSubgroup()).WillOnce(testing::Return(folly::unit));
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  // Begin a subgroup for ongoing publish activity
-  auto pubConsumer = doPublish(
-      publisherSession,
-      FullTrackName{kTestNamespace, "track_stream"},
-      /*addToState=*/false
-  );
-  // TODO: bug subscriber not added until next loop?
-  exec_->drive();
-  auto subgroupRes = pubConsumer->beginSubgroup(0, 0, 0);
-  EXPECT_TRUE(subgroupRes.hasValue());
-  auto subgroup = *subgroupRes;
-
-  // publisher ends subscription
-  EXPECT_CALL(*mockConsumer1, publishDone(testing::_));
-  EXPECT_TRUE(
-      pubConsumer->publishDone({RequestID(1), PublishDoneStatusCode::TRACK_ENDED, 0, "track ended"})
-          .hasValue()
-  );
-  subgroup->endOfSubgroup();
-
-  // Subscriber 2 subscribes to publishNamespaces but doesn't get finished track
-  doSubscribeNamespace(subscriber2, kTestNamespace);
-
-  // First publish (existing context handles initial publish), now publish a
-  // second track
-  // Expect publish calls on both subscribers, just fail them.
-  EXPECT_CALL(*subscriber1, publish(testing::_, testing::_))
-      .WillOnce([](const auto& /*pubReq*/, auto /*subHandle*/) {
-        return folly::makeUnexpected(PublishError{});
-      });
-
-  EXPECT_CALL(*subscriber2, publish(testing::_, testing::_))
-      .WillOnce([](const auto& /*pubReq*/, auto /*subHandle*/) {
-        return folly::makeUnexpected(PublishError{});
-      });
-
-  auto pubConsumer2 = doPublish(publisherSession, FullTrackName{kTestNamespace, "track_stream_2"});
-  exec_->drive();
-
-  removeSession(publisherSession);
-  removeSession(subscriber1);
-  removeSession(subscriber2);
-}
-
-// Test: Data operations return CANCELLED when all subscribers fail and are
-// removed. This verifies that when a data operation (objectPayload with
-// fin=false) causes all subscribers to error out, the operation correctly
-// detects that data went nowhere and returns CANCELLED.
-TEST_F(MoQRelayTest, DataOperationCancelledWhenAllSubscribersFail) {
-  auto publisherSession = createMockSession();
-  std::array<std::shared_ptr<MoQSession>, 2> subscribers;
-  std::array<std::shared_ptr<MockTrackConsumer>, 2> consumers;
-  std::array<std::shared_ptr<MockSubgroupConsumer>, 2> sgs;
-
-  for (auto& s : subscribers) {
-    s = createMockSession();
-  }
-  for (auto& c : consumers) {
-    c = createMockConsumer();
-  }
-
-  // Setup both subscribers to succeed on first payload, fail on second
-  for (size_t i = 0; i < 2; ++i) {
-    EXPECT_CALL(*consumers[i], beginSubgroup(0, 0, _, _))
-        .WillOnce([this, i, &sgs](uint64_t, uint64_t, uint8_t, bool) {
-          sgs[i] = createMockSubgroupConsumer();
-          EXPECT_CALL(*sgs[i], objectPayload(_, false))
-              .WillOnce(Return(folly::makeExpected<MoQPublishError>(ObjectPublishStatus::IN_PROGRESS
-              )))
-              .WillOnce(Return(folly::makeUnexpected(
-                  MoQPublishError(MoQPublishError::WRITE_ERROR, "write failed")
-              )));
-          return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sgs[i]);
-        });
-  }
-
-  // Publish track and add subscribers
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-  for (size_t i = 0; i < 2; ++i) {
-    subscribeToTrack(subscribers[i], kTestTrackName, consumers[i], RequestID(i + 1));
-  }
-
-  // Begin subgroup and object
-  auto sgRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(sgRes.hasValue());
-  auto subgroup = *sgRes;
-  EXPECT_TRUE(subgroup->beginObject(0, 100, 0).hasValue());
-
-  // First objectPayload succeeds - both subscribers receive it
-  auto res1 = subgroup->objectPayload(folly::IOBuf::copyBuffer("data1"), false);
-  EXPECT_TRUE(res1.hasValue());
-  EXPECT_EQ(*res1, ObjectPublishStatus::IN_PROGRESS);
-
-  // Second objectPayload causes both subscribers to fail and get removed
-  // This should return CANCELLED because all subscribers are gone
-  auto res2 = subgroup->objectPayload(folly::IOBuf::copyBuffer("data2"), false);
-  EXPECT_FALSE(res2.hasValue());
-  EXPECT_EQ(res2.error().code, MoQPublishError::CANCELLED);
-
-  // Clean up
-  subgroup->reset(ResetStreamErrorCode::INTERNAL_ERROR);
-  removeSession(publisherSession);
-  for (auto& s : subscribers) {
-    removeSession(s);
-  }
-}
-
-// Test: Partial subscriber failure does not cancel data operations.
-// This verifies that when some subscribers fail but others succeed, the
-// operation continues successfully and only returns CANCELLED when ALL
-// subscribers are gone.
-TEST_F(MoQRelayTest, PartialSubscriberFailureDoesNotCancelData) {
-  auto publisherSession = createMockSession();
-  std::array<std::shared_ptr<MoQSession>, 3> subscribers;
-  std::array<std::shared_ptr<MockTrackConsumer>, 3> consumers;
-  std::array<std::shared_ptr<MockSubgroupConsumer>, 3> sgs;
-
-  for (auto& s : subscribers) {
-    s = createMockSession();
-  }
-  for (auto& c : consumers) {
-    c = createMockConsumer();
-  }
-
-  // Setup subscriber 0 - will fail on second objectPayload
-  EXPECT_CALL(*consumers[0], beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sgs](uint64_t, uint64_t, uint8_t, bool) {
-        sgs[0] = createMockSubgroupConsumer();
-        EXPECT_CALL(*sgs[0], objectPayload(_, false))
-            .WillOnce(Return(folly::makeExpected<MoQPublishError>(ObjectPublishStatus::IN_PROGRESS))
-            )
-            .WillOnce(Return(
-                folly::makeUnexpected(MoQPublishError(MoQPublishError::WRITE_ERROR, "write failed"))
-            ));
-        // Expect reset when subscriber is removed
-        EXPECT_CALL(*sgs[0], reset(_)).Times(1);
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sgs[0]);
-      });
-
-  // Setup subscribers 1 and 2 - will succeed on objectPayload
-  // They will get reset during cleanup when the test calls subgroup->reset()
-  for (size_t i = 1; i < 3; ++i) {
-    EXPECT_CALL(*consumers[i], beginSubgroup(0, 0, _, _))
-        .WillOnce([this, i, &sgs](uint64_t, uint64_t, uint8_t, bool) {
-          sgs[i] = createMockSubgroupConsumer();
-          EXPECT_CALL(*sgs[i], objectPayload(_, false))
-              .WillRepeatedly(
-                  Return(folly::makeExpected<MoQPublishError>(ObjectPublishStatus::IN_PROGRESS))
-              );
-          // Will be reset by cleanup code at end of test
-          EXPECT_CALL(*sgs[i], reset(ResetStreamErrorCode::SESSION_CLOSED)).Times(1);
-          return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sgs[i]);
-        });
-  }
-
-  // Publish track and add subscribers
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-  for (size_t i = 0; i < 3; ++i) {
-    subscribeToTrack(subscribers[i], kTestTrackName, consumers[i], RequestID(i + 1));
-  }
-
-  // Begin subgroup and object
-  auto sgRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(sgRes.hasValue());
-  auto subgroup = *sgRes;
-  EXPECT_TRUE(subgroup->beginObject(0, 100, 0).hasValue());
-
-  // First objectPayload succeeds - all 3 subscribers receive it
-  auto res1 = subgroup->objectPayload(folly::IOBuf::copyBuffer("data1"), false);
-  EXPECT_TRUE(res1.hasValue());
-  EXPECT_EQ(*res1, ObjectPublishStatus::IN_PROGRESS);
-
-  // Second objectPayload causes subscriber 0 to fail and get removed
-  // But subscribers 1 and 2 still exist, so operation succeeds
-  auto res2 = subgroup->objectPayload(folly::IOBuf::copyBuffer("data2"), false);
-  EXPECT_TRUE(res2.hasValue());
-  EXPECT_EQ(*res2, ObjectPublishStatus::IN_PROGRESS);
-
-  // Clean up - reset the subgroup which resets remaining subscribers 1 and 2
-  subgroup->reset(ResetStreamErrorCode::SESSION_CLOSED);
-  removeSession(publisherSession);
-  for (auto& s : subscribers) {
-    removeSession(s);
-  }
-}
-
-// Test: SubscribeUpdate can decrease start location
-// Per spec, start location can be decreased. The subscriber should use FETCH
-// to retrieve objects between the new start and the current largest location.
-TEST_F(MoQRelayTest, SubscribeUpdateStartLocationCanDecrease) {
-  auto publisherSession = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Subscribe with initial start location
-  auto consumer = createMockConsumer();
-  SubscribeRequest sub;
-  sub.fullTrackName = kTestTrackName;
-  sub.requestID = RequestID(1);
-  sub.locType = LocationType::AbsoluteStart;
-  sub.start = AbsoluteLocation{10, 0};
-  sub.endGroup = 0; // Open-ended
-
-  std::shared_ptr<SubscriptionHandle> handle{nullptr};
-  withSessionContext(subscriberSession, [&]() {
-    auto task = relay_->subscribe(std::move(sub), consumer);
-    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-    ASSERT_TRUE(res.hasValue());
-    handle = *res;
-  });
-
-  // Cast to MoQForwarder::Subscriber to access range
-  auto* subscriber = dynamic_cast<MoQForwarder::Subscriber*>(handle.get());
-  ASSERT_NE(subscriber, nullptr);
-
-  // Verify initial start location
-  EXPECT_EQ(subscriber->range.start, (AbsoluteLocation{10, 0}));
-
-  // Send SubscribeUpdate with decreased start location
-  SubscribeUpdate subscribeUpdate{
-      RequestID(2),           // requestID for the update
-      sub.requestID,          // subscriptionRequestID
-      AbsoluteLocation{5, 0}, // Start decreased from {10, 0} to {5, 0}
-      0,                      // endGroup (open-ended)
-      kDefaultPriority,
-      true
-  }; // forward
-
-  auto updateRes = folly::coro::blockingWait(subscriber->requestUpdate(subscribeUpdate));
-  EXPECT_TRUE(updateRes.hasValue()) << "RequestUpdate with decreased start should succeed";
-
-  // Verify start location was updated
-  EXPECT_EQ(subscriber->range.start, (AbsoluteLocation{5, 0}))
-      << "Start location should be updated to {5, 0}";
-
-  removeSession(publisherSession);
-  removeSession(subscriberSession);
-}
-
-// Test: Subgroup is tombstoned after CANCELLED error (soft error)
-// Verifies that a CANCELLED error on a subgroup sets it to nullptr (tombstone)
-// but keeps the subscription alive
-TEST_F(MoQRelayTest, SubgroupTombstonedAfterCancelledError) {
-  auto publisherSession = createMockSession();
-  auto subscriber = createMockSession();
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Create subscriber with a mock consumer that will fail with CANCELLED
-  auto consumer = createMockConsumer();
-  std::shared_ptr<MockSubgroupConsumer> sg;
-
-  EXPECT_CALL(*consumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg](uint64_t, uint64_t, uint8_t, bool) {
-        sg = createMockSubgroupConsumer();
-        // First object succeeds, second fails with CANCELLED
-        EXPECT_CALL(*sg, object(0, _, _, false)).WillOnce(Return(folly::unit));
-        EXPECT_CALL(*sg, object(1, _, _, false))
-            .WillOnce(Return(
-                folly::makeUnexpected(MoQPublishError(MoQPublishError::CANCELLED, "STOP_SENDING"))
-            ));
-        // Per API contract, error implies implicit reset - no reset() call
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  // Second subgroup should still be created (subscription is alive)
-  std::shared_ptr<MockSubgroupConsumer> sg2;
-  EXPECT_CALL(*consumer, beginSubgroup(1, 0, _, _))
-      .WillOnce([this, &sg2](uint64_t, uint64_t, uint8_t, bool) {
-        sg2 = createMockSubgroupConsumer();
-        EXPECT_CALL(*sg2, endOfSubgroup()).WillOnce(Return(folly::unit));
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg2);
-      });
-
-  subscribeToTrack(subscriber, kTestTrackName, consumer, RequestID(1));
-
-  // Begin first subgroup and send objects
-  auto subgroup1Res = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroup1Res.hasValue());
-  auto subgroup1 = *subgroup1Res;
-
-  // First object succeeds
-  EXPECT_TRUE(subgroup1->object(0, test::makeBuf(10)).hasValue());
-
-  // Second object triggers CANCELLED error - subgroup gets tombstoned
-  EXPECT_TRUE(subgroup1->object(1, test::makeBuf(10)).hasValue());
-
-  // Begin second subgroup - should succeed because subscription is still alive
-  auto subgroup2Res = publishConsumer->beginSubgroup(1, 0, 0);
-  ASSERT_TRUE(subgroup2Res.hasValue());
-  auto subgroup2 = *subgroup2Res;
-
-  // End second subgroup
-  EXPECT_TRUE(subgroup2->endOfSubgroup().hasValue());
-
-  // End first subgroup (after tombstoning)
-  EXPECT_TRUE(subgroup1->endOfSubgroup().hasValue());
-
-  removeSession(publisherSession);
-  removeSession(subscriber);
-}
-
-// Test: Objects published to tombstoned subgroup are skipped for that
-// subscriber
-TEST_F(MoQRelayTest, TombstonedSubgroupIgnoresSubsequentObjects) {
-  auto publisherSession = createMockSession();
-  auto subscriber = createMockSession();
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Create subscriber that fails with CANCELLED on first object
-  auto consumer = createMockConsumer();
-  std::shared_ptr<MockSubgroupConsumer> sg;
-
-  EXPECT_CALL(*consumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg](uint64_t, uint64_t, uint8_t, bool) {
-        sg = createMockSubgroupConsumer();
-        // First object fails with CANCELLED
-        EXPECT_CALL(*sg, object(0, _, _, false))
-            .WillOnce(Return(folly::makeUnexpected(
-                MoQPublishError(MoQPublishError::CANCELLED, "delivery timeout")
-            )));
-        // Per API contract, error implies implicit reset - no reset() call
-        // After tombstoning, should NOT receive any more objects
-        EXPECT_CALL(*sg, object(1, _, _, _)).Times(0);
-        EXPECT_CALL(*sg, object(2, _, _, _)).Times(0);
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  subscribeToTrack(subscriber, kTestTrackName, consumer, RequestID(1));
-
-  // Begin subgroup
-  auto subgroupRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroupRes.hasValue());
-  auto subgroup = *subgroupRes;
-
-  // First object triggers CANCELLED - subgroup gets tombstoned
-  EXPECT_TRUE(subgroup->object(0, test::makeBuf(10)).hasValue());
-
-  // Subsequent objects should be skipped for this subscriber (tombstoned)
-  // These return CANCELLED because there are no more active subscribers
-  auto res1 = subgroup->object(1, test::makeBuf(10));
-  EXPECT_FALSE(res1.hasValue());
-  EXPECT_EQ(res1.error().code, MoQPublishError::CANCELLED);
-
-  auto res2 = subgroup->object(2, test::makeBuf(10));
-  EXPECT_FALSE(res2.hasValue());
-  EXPECT_EQ(res2.error().code, MoQPublishError::CANCELLED);
-
-  subgroup->reset(ResetStreamErrorCode::SESSION_CLOSED);
-  removeSession(publisherSession);
-  removeSession(subscriber);
-}
-
-// Test: Late joiner gets subgroup even after another subscriber tombstoned it
-TEST_F(MoQRelayTest, LateJoinerGetsSubgroupAfterTombstone) {
-  auto publisherSession = createMockSession();
-  auto subscriber1 = createMockSession();
-  auto subscriber2 = createMockSession();
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // First subscriber fails with CANCELLED (tombstones subgroup)
-  auto consumer1 = createMockConsumer();
-  std::shared_ptr<MockSubgroupConsumer> sg1;
-
-  EXPECT_CALL(*consumer1, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg1](uint64_t, uint64_t, uint8_t, bool) {
-        sg1 = createMockSubgroupConsumer();
-        EXPECT_CALL(*sg1, object(0, _, _, false))
-            .WillOnce(Return(
-                folly::makeUnexpected(MoQPublishError(MoQPublishError::CANCELLED, "STOP_SENDING"))
-            ));
-        // Per API contract, error implies implicit reset - no reset() call
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg1);
-      });
-
-  // Second subscriber (late joiner) should still get the subgroup
-  auto consumer2 = createMockConsumer();
-  std::shared_ptr<MockSubgroupConsumer> sg2;
-
-  EXPECT_CALL(*consumer2, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg2](uint64_t, uint64_t, uint8_t, bool) {
-        sg2 = createMockSubgroupConsumer();
-        EXPECT_CALL(*sg2, object(1, _, _, false)).WillOnce(Return(folly::unit));
-        EXPECT_CALL(*sg2, endOfSubgroup()).WillOnce(Return(folly::unit));
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg2);
-      });
-
-  // Subscribe first subscriber
-  subscribeToTrack(subscriber1, kTestTrackName, consumer1, RequestID(1));
-
-  // Begin subgroup
-  auto subgroupRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroupRes.hasValue());
-  auto subgroup = *subgroupRes;
-
-  // First object triggers CANCELLED for subscriber1 - tombstones their subgroup
-  EXPECT_TRUE(subgroup->object(0, test::makeBuf(10)).hasValue());
-
-  // Late joiner subscribes after tombstone
-  subscribeToTrack(subscriber2, kTestTrackName, consumer2, RequestID(2));
-
-  // Next object should go to subscriber2 only (subscriber1 is tombstoned)
-  EXPECT_TRUE(subgroup->object(1, test::makeBuf(10)).hasValue());
-
-  // End subgroup
-  EXPECT_TRUE(subgroup->endOfSubgroup().hasValue());
-
-  removeSession(publisherSession);
-  removeSession(subscriber1);
-  removeSession(subscriber2);
-}
-
-// Test: Hard errors (WRITE_ERROR) remove the entire subscription
-TEST_F(MoQRelayTest, HardErrorsRemoveSubscriber) {
-  auto publisherSession = createMockSession();
-  auto subscriber = createMockSession();
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Create subscriber that fails with WRITE_ERROR (hard error)
-  auto consumer = createMockConsumer();
-  std::shared_ptr<MockSubgroupConsumer> sg;
-
-  EXPECT_CALL(*consumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg](uint64_t, uint64_t, uint8_t, bool) {
-        sg = createMockSubgroupConsumer();
-        EXPECT_CALL(*sg, object(0, _, _, false))
-            .WillOnce(Return(folly::makeUnexpected(
-                MoQPublishError(MoQPublishError::WRITE_ERROR, "transport broken")
-            )));
-        // Should be reset when subscriber is removed
-        EXPECT_CALL(*sg, reset(_)).Times(1);
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  // publishDone should be called because subscription is being removed
-  EXPECT_CALL(*consumer, publishDone(_))
-      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-
-  // Second subgroup should NOT be created (subscriber removed)
-  EXPECT_CALL(*consumer, beginSubgroup(1, _, _, _)).Times(0);
-
-  subscribeToTrack(subscriber, kTestTrackName, consumer, RequestID(1));
-
-  // Begin first subgroup
-  auto subgroup1Res = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroup1Res.hasValue());
-  auto subgroup1 = *subgroup1Res;
-
-  // Object triggers WRITE_ERROR - subscriber removed entirely
-  // The call returns CANCELLED because no subscriber received the object
-  auto res = subgroup1->object(0, test::makeBuf(10));
-  EXPECT_FALSE(res.hasValue());
-  EXPECT_EQ(res.error().code, MoQPublishError::CANCELLED);
-
-  // Try to begin second subgroup - also returns CANCELLED (no subscribers)
-  auto subgroup2Res = publishConsumer->beginSubgroup(1, 0, 0);
-  EXPECT_FALSE(subgroup2Res.hasValue());
-  EXPECT_EQ(subgroup2Res.error().code, MoQPublishError::CANCELLED);
-
-  subgroup1->reset(ResetStreamErrorCode::SESSION_CLOSED);
-  removeSession(publisherSession);
-  removeSession(subscriber);
-}
-
-// Regression test: endOfSubgroup with hard error used to cause use-after-free.
-// forEachSubscriber held a const reference to the subscriber's shared_ptr in
-// the map. When endOfSubgroup returned a hard error, handleSubgroupError
-// removed the subscriber from the map (destroying the shared_ptr). The onError
-// lambda's captured copy was then the last reference. After it was destroyed,
-// closeSubgroupForSubscriber accessed the freed subscriber.
-TEST_F(MoQRelayTest, EndOfSubgroupHardErrorDoesNotCrash) {
-  auto publisherSession = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  // Setup: Publish track
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Create subscriber that fails with WRITE_ERROR on endOfSubgroup
-  auto consumer = createMockConsumer();
-  std::shared_ptr<MockSubgroupConsumer> sg;
-
-  EXPECT_CALL(*consumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg](uint64_t, uint64_t, uint8_t, bool) {
-        sg = createMockSubgroupConsumer();
-        // Override endOfSubgroup to return a hard error (WRITE_ERROR)
-        EXPECT_CALL(*sg, endOfSubgroup())
-            .WillOnce(Return(folly::makeUnexpected(
-                MoQPublishError(MoQPublishError::WRITE_ERROR, "transport broken")
-            )));
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  // publishDone is called when subscription is removed due to hard error
-  EXPECT_CALL(*consumer, publishDone(_))
-      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-
-  subscribeToTrack(
-      subscriberSession,
-      kTestTrackName,
-      consumer,
-      RequestID(1),
-      /*addToState=*/false
-  );
-
-  // Begin subgroup - sets up the subscriber's subgroup consumer
-  auto subgroupRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroupRes.hasValue());
-  auto subgroup = *subgroupRes;
-
-  // endOfSubgroup triggers the use-after-free without the fix:
-  // subscriber's endOfSubgroup() returns WRITE_ERROR -> handleSubgroupError
-  // removes subscriber from map -> onError lambda destroyed (last shared_ptr
-  // ref) -> closeSubgroupForSubscriber accesses freed subscriber
-  subgroup->endOfSubgroup();
-
-  removeSession(publisherSession);
-  removeSession(subscriberSession);
-}
-
-// Helper callback that drops the forwarder shared_ptr when onEmpty fires,
-// simulating what happens in production when the last reference holder
-// (e.g., SubscribeWriteback destructor chain) destroys the forwarder.
-class ForwarderDestroyingCallback : public MoQForwarder::Callback {
-public:
-  explicit ForwarderDestroyingCallback(std::shared_ptr<MoQForwarder>& forwarderRef)
-      : forwarderRef_(forwarderRef) {}
-
-  void onEmpty(MoQForwarder*) override {
-    // Drop the last shared_ptr, destroying the forwarder.
-    // This simulates the production scenario where onEmpty ->
-    // subscriptions_.erase -> MoQForwarder destructor.
-    forwarderRef_.reset();
-  }
-
-private:
-  std::shared_ptr<MoQForwarder>& forwarderRef_;
-};
-
-TEST_F(MoQRelayTest, SubscribeNamespaceEmptyPrefixRejectedPreV16) {
-  // Default session uses kVersionDraftCurrent (draft-14, which is < 16)
-  auto session = createMockSession();
-
-  TrackNamespace emptyNs{{}};
-  SubscribeNamespace subNs;
-  subNs.trackNamespacePrefix = emptyNs;
-
-  withSessionContext(session, [&]() {
-    auto task = relay_->subscribeNamespace(std::move(subNs), nullptr);
-    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-    ASSERT_FALSE(res.hasValue()
-    ) << "Empty namespace prefix should be rejected for pre-v16 sessions";
-    EXPECT_EQ(res.error().errorCode, SubscribeNamespaceErrorCode::NAMESPACE_PREFIX_UNKNOWN);
-    EXPECT_EQ(res.error().reasonPhrase, "empty");
-  });
-
-  removeSession(session);
-}
-
-TEST_F(MoQRelayTest, SubscribeNamespaceEmptyPrefixAllowedV16) {
-  auto session = createMockSession();
-  // Override the negotiated version to draft-16
-  ON_CALL(*session, getNegotiatedVersion())
-      .WillByDefault(Return(std::optional<uint64_t>(kVersionDraft16)));
-
-  TrackNamespace emptyNs{{}};
-  doSubscribeNamespace(session, emptyNs);
-
-  removeSession(session);
-}
-
-// Regression test: SubgroupForwarder::reset() use-after-free.
-// When the forwarder is draining (publishDone received) and a subscriber's
-// last subgroup is closed via reset(), closeSubgroupForSubscriber calls
-// removeSubscriber -> checkAndFireOnEmpty -> onEmpty callback. The onEmpty
-// callback destroys the MoQForwarder (by dropping the last shared_ptr).
-// But reset() is still inside forEachSubscriberSubgroup, which accesses
-// forwarder_.subscribers_ after the lambda returns, causing
-// heap-use-after-free.
-TEST_F(MoQRelayTest, ResetDuringDrainingDoesNotCrash) {
-  auto session = createMockSession();
-
-  auto forwarder = std::make_shared<MoQForwarder>(kTestTrackName, AbsoluteLocation{0, 0});
-  auto callback = std::make_shared<ForwarderDestroyingCallback>(forwarder);
-  forwarder->setCallback(callback);
-
-  // Create a subscriber
-  auto consumer = createMockConsumer();
-  std::shared_ptr<MockSubgroupConsumer> sg;
-
-  EXPECT_CALL(*consumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg](uint64_t, uint64_t, uint8_t, bool) {
-        sg = createMockSubgroupConsumer();
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-
-  EXPECT_CALL(*consumer, publishDone(_))
-      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-
-  SubscribeRequest sub;
-  sub.fullTrackName = kTestTrackName;
-  sub.requestID = RequestID(1);
-  sub.locType = LocationType::LargestGroup;
-  forwarder->addSubscriber(session, sub, consumer);
-
-  // Begin subgroup
-  auto subgroupRes = forwarder->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroupRes.hasValue());
-  auto subgroup = *subgroupRes;
-
-  // Drain the subscriber
-  forwarder->publishDone(
-      PublishDone{RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher ended"}
-  );
-
-  // Now reset the subgroup. This triggers:
-  //  SubgroupForwarder::reset()
-  //    -> forEachSubscriberSubgroup()
-  //      -> reset subscriber's subgroupConsumer
-  //      -> closeSubgroupForSubscriber()
-  //        -> sub->shouldRemove() returns true (draining, last subgroup)
-  //        -> removeSubscriber()
-  //          -> checkAndFireOnEmpty()
-  //            -> onEmpty destroys the forwarder (drops shared_ptr)
-  //    -> forEachSubscriberSubgroup accesses forwarder_.subscribers_ -> CRASH
-  EXPECT_CALL(*sg, reset(_)).Times(1);
-  subgroup->reset(ResetStreamErrorCode::INTERNAL_ERROR);
-
-  // forwarder should have been destroyed by onEmpty
-  EXPECT_EQ(forwarder, nullptr);
-}
-
-// Same as above but with multiple subscribers draining.
-TEST_F(MoQRelayTest, ResetDuringDrainingMultipleSubscribersDoesNotCrash) {
-  auto session1 = createMockSession();
-  auto session2 = createMockSession();
-
-  auto forwarder = std::make_shared<MoQForwarder>(kTestTrackName, AbsoluteLocation{0, 0});
-  auto callback = std::make_shared<ForwarderDestroyingCallback>(forwarder);
-  forwarder->setCallback(callback);
-
-  // Create two subscribers
-  std::shared_ptr<MockSubgroupConsumer> sg1, sg2;
-  auto consumer1 = createMockConsumer();
-  auto consumer2 = createMockConsumer();
-
-  EXPECT_CALL(*consumer1, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg1](uint64_t, uint64_t, uint8_t, bool) {
-        sg1 = createMockSubgroupConsumer();
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg1);
-      });
-
-  EXPECT_CALL(*consumer2, beginSubgroup(0, 0, _, _))
-      .WillOnce([this, &sg2](uint64_t, uint64_t, uint8_t, bool) {
-        sg2 = createMockSubgroupConsumer();
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg2);
-      });
-
-  EXPECT_CALL(*consumer1, publishDone(_))
-      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-  EXPECT_CALL(*consumer2, publishDone(_))
-      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-
-  SubscribeRequest sub1;
-  sub1.fullTrackName = kTestTrackName;
-  sub1.requestID = RequestID(1);
-  sub1.locType = LocationType::LargestGroup;
-  forwarder->addSubscriber(session1, sub1, consumer1);
-
-  SubscribeRequest sub2;
-  sub2.fullTrackName = kTestTrackName;
-  sub2.requestID = RequestID(2);
-  sub2.locType = LocationType::LargestGroup;
-  forwarder->addSubscriber(session2, sub2, consumer2);
-
-  // Begin subgroup - both subscribers get it
-  auto subgroupRes = forwarder->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroupRes.hasValue());
-  auto subgroup = *subgroupRes;
-
-  // Drain both subscribers
-  forwarder->publishDone(
-      PublishDone{RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher ended"}
-  );
-
-  // Reset the subgroup. Both subscribers are draining with only this subgroup.
-  // As each is removed, the second removal triggers onEmpty which destroys
-  // the forwarder while we may still be iterating.
-  EXPECT_CALL(*sg1, reset(_)).Times(1);
-  EXPECT_CALL(*sg2, reset(_)).Times(1);
-  subgroup->reset(ResetStreamErrorCode::INTERNAL_ERROR);
-
-  EXPECT_EQ(forwarder, nullptr);
-}
-
-// ============================================================
-// Extensions Tests
-// ============================================================
-
-// Test: Extensions from publish are forwarded to subscribers via
-// subscribeNamespace
-TEST_F(MoQRelayTest, PublishExtensionsForwardedToSubscribers) {
-  auto publisherSession = createMockSession();
-  auto subscriber = createMockSession();
-
-  // Subscribe to namespace first
-  auto mockConsumer = createMockConsumer();
-  Extensions receivedExtensions;
-  EXPECT_CALL(*subscriber, publish(testing::_, testing::_))
-      .WillOnce([&mockConsumer,
-                 &receivedExtensions](const PublishRequest& pubReq, auto /*subHandle*/) {
-        receivedExtensions = pubReq.extensions;
-        return Subscriber::PublishResult(Subscriber::PublishConsumerAndReplyTask{
-            mockConsumer,
-            []() -> folly::coro::Task<folly::Expected<PublishOk, PublishError>> {
-              co_return PublishOk{
-                  RequestID(1),
-                  true,
-                  0,
-                  GroupOrder::OldestFirst,
-                  LocationType::LargestObject,
-                  std::nullopt,
-                  std::nullopt
-              };
-            }()
-        });
-      });
-
-  doSubscribeNamespace(subscriber, kTestNamespace);
-
-  // Publish with extensions (both known and unknown)
-  PublishRequest pub;
-  pub.fullTrackName = kTestTrackName;
-  pub.extensions.insertMutableExtension(Extension{kDeliveryTimeoutExtensionType, 5000});
-  pub.extensions.insertMutableExtension(Extension{0xBEEF'0000, 42});
-
-  withSessionContext(publisherSession, [&]() {
-    auto res = relay_->publish(std::move(pub), createMockSubscriptionHandle());
-    EXPECT_TRUE(res.hasValue());
-    if (res.hasValue()) {
-      getOrCreateMockState(publisherSession)->publishConsumers.push_back(res->consumer);
-    }
-  });
-  exec_->drive();
-
-  // Verify extensions were forwarded
-  EXPECT_EQ(receivedExtensions.getIntExtension(kDeliveryTimeoutExtensionType), 5000);
-  EXPECT_EQ(receivedExtensions.getIntExtension(0xBEEF'0000), 42);
-
-  removeSession(publisherSession);
-  removeSession(subscriber);
-}
-
-// ============================================================
-// Dynamic Groups Extension Tests
-// ============================================================
-
-// Test: relay PUBLISH path – dynamic groups from PublishRequest extensions
-// is stored in the forwarder and forwarded to every downstream subscriber
-TEST_F(MoQRelayTest, RelayPublishPropagatesDynamicGroupsToSubscribers) {
-  auto publisherSession = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  // Build a PublishRequest with DYNAMIC_GROUPS enabled
-  PublishRequest pub;
-  pub.fullTrackName = kTestTrackName;
-  setPublisherDynamicGroups(pub, true);
-
-  withSessionContext(publisherSession, [&]() {
-    auto res = relay_->publish(std::move(pub), createMockSubscriptionHandle());
-    ASSERT_TRUE(res.hasValue());
-    getOrCreateMockState(publisherSession)->publishConsumers.push_back(res->consumer);
-  });
-
-  auto consumer = createMockConsumer();
-  auto handle = subscribeToTrack(subscriberSession, kTestTrackName, consumer, RequestID(1));
-  ASSERT_NE(handle, nullptr);
-
-  auto dynGroups = getPublisherDynamicGroups(handle->subscribeOk());
-  ASSERT_TRUE(dynGroups.has_value());
-  EXPECT_TRUE(*dynGroups);
-
-  removeSession(subscriberSession);
-  exec_->drive();
-  removeSession(publisherSession);
-}
-
-// Test: Extensions from publish are forwarded to late-joining subscribers
-TEST_F(MoQRelayTest, PublishExtensionsForwardedToLateJoiners) {
-  auto publisherSession = createMockSession();
-  auto subscriber1 = createMockSession();
-  auto subscriber2 = createMockSession();
-
-  // Subscriber 1 subscribes first
-  auto mockConsumer1 = createMockConsumer();
-  EXPECT_CALL(*subscriber1, publish(testing::_, testing::_))
-      .WillOnce([&mockConsumer1](const auto&, auto) {
-        return Subscriber::PublishResult(Subscriber::PublishConsumerAndReplyTask{
-            mockConsumer1,
-            []() -> folly::coro::Task<folly::Expected<PublishOk, PublishError>> {
-              co_return PublishOk{
-                  RequestID(1),
-                  true,
-                  0,
-                  GroupOrder::OldestFirst,
-                  LocationType::LargestObject,
-                  std::nullopt,
-                  std::nullopt
-              };
-            }()
-        });
-      });
-
-  doSubscribeNamespace(subscriber1, kTestNamespace);
-
-  // Publish with extensions
-  PublishRequest pub;
-  pub.fullTrackName = kTestTrackName;
-  pub.extensions.insertMutableExtension(Extension{kDeliveryTimeoutExtensionType, 3000});
-  pub.extensions.insertMutableExtension(Extension{0xCAFE'0000, 99});
-
-  withSessionContext(publisherSession, [&]() {
-    auto res = relay_->publish(std::move(pub), createMockSubscriptionHandle());
-    EXPECT_TRUE(res.hasValue());
-    if (res.hasValue()) {
-      getOrCreateMockState(publisherSession)->publishConsumers.push_back(res->consumer);
-    }
-  });
-  exec_->drive();
-
-  // Late-joining subscriber 2 should also get extensions
-  Extensions receivedExtensions;
-  auto mockConsumer2 = createMockConsumer();
-  EXPECT_CALL(*subscriber2, publish(testing::_, testing::_))
-      .WillOnce([&mockConsumer2, &receivedExtensions](const PublishRequest& pubReq, auto) {
-        receivedExtensions = pubReq.extensions;
-        return Subscriber::PublishResult(Subscriber::PublishConsumerAndReplyTask{
-            mockConsumer2,
-            []() -> folly::coro::Task<folly::Expected<PublishOk, PublishError>> {
-              co_return PublishOk{
-                  RequestID(2),
-                  true,
-                  0,
-                  GroupOrder::OldestFirst,
-                  LocationType::LargestObject,
-                  std::nullopt,
-                  std::nullopt
-              };
-            }()
-        });
-      });
-
-  doSubscribeNamespace(subscriber2, kTestNamespace);
-  exec_->drive();
-
-  // Verify late-joiner received extensions
-  EXPECT_EQ(receivedExtensions.getIntExtension(kDeliveryTimeoutExtensionType), 3000);
-  EXPECT_EQ(receivedExtensions.getIntExtension(0xCAFE'0000), 99);
-
-  removeSession(publisherSession);
-  removeSession(subscriber1);
-  removeSession(subscriber2);
-}
-
-// Test: relay SUBSCRIBE path – dynamic groups from the upstream SubscribeOk is
-// stored in the forwarder and forwarded to both the first and late-joining
-// downstream subscribers
-TEST_F(MoQRelayTest, RelaySubscribePropagatesDynamicGroupsToAllSubscribers) {
-  auto publisherSession = createMockSession();
-  auto subscriber1 = createMockSession();
-  auto subscriber2 = createMockSession();
-
-  doPublishNamespace(publisherSession, kTestNamespace);
-
-  // Upstream returns a SubscribeOk with DYNAMIC_GROUPS = true
-  SubscribeOk upstreamOk;
-  upstreamOk.requestID = RequestID(1);
-  upstreamOk.trackAlias = TrackAlias(1);
-  upstreamOk.expires = std::chrono::milliseconds(0);
-  upstreamOk.groupOrder = GroupOrder::OldestFirst;
-  setPublisherDynamicGroups(upstreamOk, true);
-
-  EXPECT_CALL(*publisherSession, subscribe(_, _))
-      .WillOnce([upstreamOk](const auto& /*req*/, auto /*consumer*/) {
-        auto handle = std::make_shared<NiceMock<MockSubscriptionHandle>>(upstreamOk);
-        return folly::coro::makeTask<Publisher::SubscribeResult>(
-            folly::Expected<std::shared_ptr<SubscriptionHandle>, SubscribeError>(handle)
-        );
-      });
-
-  // First subscriber
-  auto consumer1 = createMockConsumer();
-  auto handle1 = subscribeToTrack(subscriber1, kTestTrackName, consumer1, RequestID(1));
-  ASSERT_NE(handle1, nullptr);
-  auto dynGroups1 = getPublisherDynamicGroups(handle1->subscribeOk());
-  ASSERT_TRUE(dynGroups1.has_value());
-  EXPECT_TRUE(*dynGroups1);
-
-  // Late-joining second subscriber – forwarder should propagate the stored
-  // dynamic groups value without another upstream roundtrip
-  auto consumer2 = createMockConsumer();
-  auto handle2 = subscribeToTrack(subscriber2, kTestTrackName, consumer2, RequestID(2));
-  ASSERT_NE(handle2, nullptr);
-  auto dynGroups2 = getPublisherDynamicGroups(handle2->subscribeOk());
-  ASSERT_TRUE(dynGroups2.has_value());
-  EXPECT_TRUE(*dynGroups2);
-
-  removeSession(publisherSession);
-  removeSession(subscriber1);
-  removeSession(subscriber2);
-}
-
-// Test: Extensions are included in SubscribeOk for subscribe path subscribers
-TEST_F(MoQRelayTest, ExtensionsIncludedInSubscribeOkForSubscribers) {
-  auto session1 = createMockSession();
-  auto session2 = createMockSession();
-
-  // Create forwarder with extensions
-  auto forwarder = std::make_shared<MoQForwarder>(kTestTrackName, AbsoluteLocation{0, 0});
-  Extensions ext;
-  ext.insertMutableExtension(Extension{kDeliveryTimeoutExtensionType, 7000});
-  ext.insertMutableExtension(Extension{0xDEAD'0000, 123});
-  forwarder->setExtensions(ext);
-
-  // Add subscriber via SubscribeRequest
-  auto consumer1 = createMockConsumer();
-  SubscribeRequest sub1;
-  sub1.fullTrackName = kTestTrackName;
-  sub1.requestID = RequestID(1);
-  sub1.locType = LocationType::LargestGroup;
-  auto subscriber1 = forwarder->addSubscriber(session1, sub1, consumer1);
-  ASSERT_NE(subscriber1, nullptr);
-
-  // Verify extensions are in the SubscribeOk
-  EXPECT_EQ(
-      subscriber1->subscribeOk().extensions.getIntExtension(kDeliveryTimeoutExtensionType),
-      7000
-  );
-  EXPECT_EQ(subscriber1->subscribeOk().extensions.getIntExtension(0xDEAD'0000), 123);
-
-  // Second subscriber should also get extensions
-  auto consumer2 = createMockConsumer();
-  SubscribeRequest sub2;
-  sub2.fullTrackName = kTestTrackName;
-  sub2.requestID = RequestID(2);
-  sub2.locType = LocationType::LargestGroup;
-  auto subscriber2 = forwarder->addSubscriber(session2, sub2, consumer2);
-  ASSERT_NE(subscriber2, nullptr);
-
-  EXPECT_EQ(
-      subscriber2->subscribeOk().extensions.getIntExtension(kDeliveryTimeoutExtensionType),
-      7000
-  );
-  EXPECT_EQ(subscriber2->subscribeOk().extensions.getIntExtension(0xDEAD'0000), 123);
-
-  // Add subscriber via publish path (bool forward)
-  auto session3 = createMockSession();
-  auto subscriber3 = forwarder->addSubscriber(session3, /*forward=*/false);
-  ASSERT_NE(subscriber3, nullptr);
-
-  EXPECT_EQ(
-      subscriber3->subscribeOk().extensions.getIntExtension(kDeliveryTimeoutExtensionType),
-      7000
-  );
-  EXPECT_EQ(subscriber3->subscribeOk().extensions.getIntExtension(0xDEAD'0000), 123);
-}
-
-TEST_F(MoQRelayTest, ExactNamespaceSubscriberReceivesPublishNamespace) {
-  auto subscriber = createMockSession();
-  auto publisher = createMockSession();
-
-  // Subscriber subscribes to exact namespace {"test", "namespace"}
-  doSubscribeNamespace(subscriber, kTestNamespace);
-
-  // Expect the subscriber to receive a publishNamespace forwarding when
-  // the publisher announces the same exact namespace
-  EXPECT_CALL(*subscriber, publishNamespace(_, _))
-      .WillOnce(
-          [](PublishNamespace ann, auto) -> folly::coro::Task<Subscriber::PublishNamespaceResult> {
-            EXPECT_EQ(ann.trackNamespace, kTestNamespace);
-            co_return folly::makeUnexpected(PublishNamespaceError{
-                ann.requestID,
-                PublishNamespaceErrorCode::UNINTERESTED,
-                "test"
-            });
-          }
-      );
-
-  // Publisher announces the same exact namespace
-  doPublishNamespace(publisher, kTestNamespace);
-
-  // Drive the executor so the async publishNamespace forwarding runs
-  exec_->drive();
-
-  removeSession(publisher);
-  removeSession(subscriber);
-}
-
-// Test: TrackStatus on non-existent track
-TEST_F(MoQRelayTest, TrackStatusNonExistentTrack) {
-  auto clientSession = createMockSession();
-
-  // Request trackStatus for a track that doesn't exist
-  TrackStatus trackStatus;
-  trackStatus.fullTrackName = kTestTrackName;
-  trackStatus.requestID = RequestID(1);
-
-  withSessionContext(clientSession, [&]() {
-    auto task = relay_->trackStatus(trackStatus);
-    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-
-    // Should return error indicating track not found
-    EXPECT_FALSE(res.hasValue());
-    EXPECT_EQ(res.error().errorCode, TrackStatusErrorCode::TRACK_NOT_EXIST);
-    EXPECT_FALSE(res.error().reasonPhrase.empty());
-  });
-
-  removeSession(clientSession);
-}
-
-// Test: TrackStatus on existing track - returns forwarder state (no upstream
-// call)
-TEST_F(MoQRelayTest, TrackStatusSuccessfulForward) {
-  auto publisherSession = createMockSession();
-  auto clientSession = createMockSession();
-
-  doPublish(publisherSession, kTestTrackName);
-
-  auto consumer = createMockConsumer();
-  subscribeToTrack(clientSession, kTestTrackName, consumer, RequestID(1));
-
-  TrackStatus trackStatus;
-  trackStatus.fullTrackName = kTestTrackName;
-  trackStatus.requestID = RequestID(2);
-
-  withSessionContext(clientSession, [&]() {
-    auto task = relay_->trackStatus(trackStatus);
-    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-
-    // Should return status from local forwarder
-    // Since no data was sent, statusCode should be TRACK_NOT_STARTED
-    EXPECT_TRUE(res.hasValue());
-    EXPECT_EQ(res.value().statusCode, TrackStatusCode::TRACK_NOT_STARTED);
-    EXPECT_EQ(res.value().fullTrackName, kTestTrackName);
-  });
-
-  removeSession(clientSession);
-  exec_->drive();
-  removeSession(publisherSession);
-}
-
-// Test: TrackStatus using namespace prefix matching (no exact subscription)
-// Verifies that when there's no exact subscription but a publisher has
-// published a matching namespace prefix, the relay correctly routes
-// TRACK_STATUS upstream using prefix matching
-TEST_F(MoQRelayTest, TrackStatusViaPrefixMatching) {
-  auto publisher = createMockSession();
-  auto requester = createMockSession();
-
-  // Publisher publishes namespace but NOT the specific track
-  doPublishNamespace(publisher, kTestNamespace);
-
-  // No exact subscription exists for kTestTrackName, so trackStatus should
-  // use prefix matching to find the publisher
-
-  // Mock the upstream trackStatus call
-  TrackStatusOk statusOk;
-  statusOk.requestID = RequestID(1);
-  statusOk.trackAlias = TrackAlias(0);
-  statusOk.largest = AbsoluteLocation{50, 25};
-
-  EXPECT_CALL(*publisher, trackStatus(_)).WillOnce([statusOk](const auto& /*ts*/) {
-    return folly::coro::makeTask<Publisher::TrackStatusResult>(statusOk);
-  });
-
-  // Execute trackStatus from requester's perspective
-  TrackStatus trackStatus;
-  trackStatus.requestID = RequestID(1);
-  trackStatus.fullTrackName = kTestTrackName;
-
-  withSessionContext(requester, [&]() {
-    auto task = relay_->trackStatus(trackStatus);
-    auto result = folly::coro::blockingWait(std::move(task), exec_.get());
-
-    // Should successfully forward via prefix matching and return the result
-    EXPECT_TRUE(result.hasValue()) << "TrackStatus via namespace prefix matching should succeed";
-    EXPECT_EQ(result.value().requestID, RequestID(1));
-    EXPECT_TRUE(result.value().largest.has_value());
-    EXPECT_EQ(result.value().largest->group, 50);
-    EXPECT_EQ(result.value().largest->object, 25);
-  });
-
-  removeSession(publisher);
-  removeSession(requester);
-}
-
-// Test: makeNamespaceBridgeHandle routes namespaceMsg to doPublishNamespace
-TEST_F(MoQRelayTest, NamespaceBridgeHandleForwardsNamespaceMsg) {
-  auto peerSession = createMockSession();
-
-  // Bridge handle routes NAMESPACE messages from peerSession into the relay.
-  // For peering the subscription prefix is empty, so the suffix == full namespace.
-  auto handle = makeNamespaceBridgeHandle(relay_, peerSession);
-  handle->namespaceMsg(kTestNamespace);
-
-  auto sessions = relay_->findPublishNamespaceSessions(kTestNamespace);
-  ASSERT_EQ(sessions.size(), 1u);
-  EXPECT_EQ(sessions[0], peerSession);
-
-  removeSession(peerSession);
-}
-
-// Test: makeNamespaceBridgeHandle routes namespaceDoneMsg to doPublishNamespaceDone
-TEST_F(MoQRelayTest, NamespaceBridgeHandleForwardsDoneMsg) {
-  auto peerSession = createMockSession();
-  auto handle = makeNamespaceBridgeHandle(relay_, peerSession);
-
-  handle->namespaceMsg(kTestNamespace);
-  ASSERT_EQ(relay_->findPublishNamespaceSessions(kTestNamespace).size(), 1u);
-
-  handle->namespaceDoneMsg(kTestNamespace);
-  EXPECT_TRUE(relay_->findPublishNamespaceSessions(kTestNamespace).empty());
-
-  removeSession(peerSession);
-}
-
-// Regression test: publisher reconnect after disconnect with active subscriber
-// crashes with SIGSEGV at MoqxRelay.cpp:463.
-//
-// Scenario (relay chain with downstream subscriber that has an open subgroup):
-//   1. Publisher session A publishes a track and opens a subgroup.
-//   2. Subscriber session B subscribes; the open subgroup is forwarded to B,
-//      so B has a live subgroup entry in the forwarder.
-//   3. Session A's connection breaks: publishDone fires, onPublishDone() resets
-//      handle to null.  The forwarder does NOT remove B because B still has an
-//      open subgroup (drainSubscriber marks B as receivedPublishDone_ instead).
-//      The subscriptions_ entry therefore survives with handle == nullptr.
-//   4. Session A reconnects and re-publishes the same track.  The multipublisher
-//      check finds the surviving entry and calls it->second.handle->unsubscribe()
-//      — null-pointer dereference, SIGSEGV.
-TEST_F(MoQRelayTest, PublisherReconnectWithOpenSubgroupNoSegfault) {
-  auto publisherSession = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  doPublishNamespace(publisherSession, kTestNamespace);
-
-  // Step 1: publisher session A publishes the track.
-  // Don't add consumer to state — we control cleanup manually.
-  auto consumer = doPublish(publisherSession, kTestTrackName, /*addToState=*/false);
-  ASSERT_NE(consumer, nullptr);
-
-  // Step 2: subscriber session B subscribes.  Wire its mock consumer to return
-  // a live SubgroupConsumer so the forwarder stores an open subgroup for B.
-  auto mockSubgroupConsumer = createMockSubgroupConsumer();
-  auto mockConsumer = createMockConsumer();
-  ON_CALL(*mockConsumer, beginSubgroup(_, _, _, _))
-      .WillByDefault(Return(folly::makeExpected<MoQPublishError>(
-          std::static_pointer_cast<SubgroupConsumer>(mockSubgroupConsumer)
-      )));
-  auto subHandle = subscribeToTrack(subscriberSession, kTestTrackName, mockConsumer);
-  ASSERT_NE(subHandle, nullptr);
-
-  // Open a subgroup through the publisher consumer so that B gets a live entry
-  // in its subgroups map inside the forwarder.
-  withSessionContext(publisherSession, [&]() {
-    auto sgRes = consumer->beginSubgroup(/*groupID=*/0, /*subgroupID=*/0, /*priority=*/0, false);
-    ASSERT_TRUE(sgRes.hasValue()) << "beginSubgroup should succeed";
-  });
-
-  // Step 3: session A's connection drops WITHOUT closing the subgroup.
-  // publishDone → onPublishDone (handle = null) + forwarder drainSubscriber.
-  // Because B has an open subgroup, B stays in the forwarder
-  // (receivedPublishDone_ = true) — subscriptions_ entry survives with
-  // handle == nullptr.
-  withSessionContext(publisherSession, [&]() {
-    consumer->publishDone(
-        {RequestID(0), PublishDoneStatusCode::SESSION_CLOSED, 0, "upstream disconnect"}
-    );
-  });
-
-  // Step 4: session A reconnects and re-publishes the same track.
-  // Before the fix this crashes (null handle->unsubscribe()).
-  auto reconnectedSession = createMockSession();
-  doPublishNamespace(reconnectedSession, kTestNamespace);
-  auto consumer2 = doPublish(reconnectedSession, kTestTrackName);
-  EXPECT_NE(consumer2, nullptr) << "re-publish after reconnect should succeed";
-
-  removeSession(publisherSession);
-  removeSession(subscriberSession);
-  removeSession(reconnectedSession);
-}
-
-// ============================================================
-// New Group Request (NGR) Tests
-// ============================================================
-
-namespace {
-// Simple callback that records every newGroupRequested call.
-struct TestNGRCallback : public MoQForwarder::Callback {
-  void onEmpty(MoQForwarder*) override {}
-  void newGroupRequested(MoQForwarder*, uint64_t group) override { calls.push_back(group); }
-  std::vector<uint64_t> calls;
-};
-
-// Build a minimal params object carrying NEW_GROUP_REQUEST=val.
-auto makeNGRParams(uint64_t val) {
-  RequestUpdate upd;
-  upd.params.insertParam(
-      Parameter(folly::to_underlying(TrackRequestParamKey::NEW_GROUP_REQUEST), val)
-  );
-  return upd.params;
-}
-} // namespace
-
-// Forwarder unit test: tryProcessNewGroupRequest gating and clearing logic.
-// Verifies all three gates (dynamic groups, largest visibility, outstanding
-// deduplication) and that updateLargest clears the outstanding request.
-TEST_F(MoQRelayTest, ForwarderNGRGatingAndClearingLogic) {
-  auto cb = std::make_shared<TestNGRCallback>();
-  auto forwarder = std::make_shared<MoQForwarder>(kTestTrackName);
-  forwarder->setCallback(cb);
-
-  // Gate 1: dynamic groups not enabled — never fires
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(5));
-  EXPECT_TRUE(cb->calls.empty());
-
-  // Enable dynamic groups
-  PublishRequest pub;
-  setPublisherDynamicGroups(pub, true);
-  forwarder->setExtensions(pub.extensions);
-
-  // Baseline: no largest, no outstanding — fires for group 5
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(5));
-  ASSERT_EQ(cb->calls.size(), 1u);
-  EXPECT_EQ(cb->calls[0], 5u);
-  cb->calls.clear();
-
-  // Deduplication: outstanding=5 blocks re-requesting group 5
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(5));
-  EXPECT_TRUE(cb->calls.empty());
-
-  // Gate 2: setLargest(10) blocks groups <= 10; group 11 fires (outstanding=5
-  // is superseded)
-  forwarder->setLargest({10, 0});
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(5));
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(10));
-  EXPECT_TRUE(cb->calls.empty());
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(11));
-  ASSERT_EQ(cb->calls.size(), 1u);
-  EXPECT_EQ(cb->calls[0], 11u);
-  cb->calls.clear();
-
-  // Gate 3: outstanding=11 blocks values <= 11; group 12 fires
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(11));
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(10));
-  EXPECT_TRUE(cb->calls.empty());
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(12));
-  ASSERT_EQ(cb->calls.size(), 1u);
-  EXPECT_EQ(cb->calls[0], 12u);
-  cb->calls.clear();
-
-  // updateLargest past outstanding clears it — group 14 can fire again
-  forwarder->updateLargest(13, 0);
-  forwarder->tryProcessNewGroupRequest(makeNGRParams(14));
-  ASSERT_EQ(cb->calls.size(), 1u);
-  EXPECT_EQ(cb->calls[0], 14u);
-}
-
-// Relay test: When a late-joining subscriber sends NEW_GROUP_REQUEST in its
-// SUBSCRIBE, the relay forwards it upstream via REQUEST_UPDATE
-TEST_F(MoQRelayTest, RelaySubscribeLateJoinerNGRForwardedUpstream) {
-  auto publisherSession = createMockSession();
-  auto subscriber1 = createMockSession();
-  auto subscriber2 = createMockSession();
-
-  doPublishNamespace(publisherSession, kTestNamespace);
-
-  // Upstream SubscribeOk advertises DYNAMIC_GROUPS = true
-  SubscribeOk upstreamOk;
-  upstreamOk.requestID = RequestID(100);
-  upstreamOk.trackAlias = TrackAlias(1);
-  upstreamOk.expires = std::chrono::milliseconds(0);
-  upstreamOk.groupOrder = GroupOrder::OldestFirst;
-  setPublisherDynamicGroups(upstreamOk, true);
-
-  auto upstreamHandle = std::make_shared<NiceMock<MockSubscriptionHandle>>(upstreamOk);
-  ON_CALL(*upstreamHandle, requestUpdateResult())
-      .WillByDefault(Return(folly::makeExpected<RequestError>(
-          RequestOk{RequestID(0), TrackRequestParameters(FrameType::REQUEST_OK), {}}
-      )));
-
-  EXPECT_CALL(*publisherSession, subscribe(_, _))
-      .WillOnce([&upstreamHandle](const auto& /*req*/, auto /*consumer*/) {
-        return folly::coro::makeTask<Publisher::SubscribeResult>(
-            folly::Expected<std::shared_ptr<SubscriptionHandle>, SubscribeError>(upstreamHandle)
-        );
-      });
-
-  // First subscriber establishes the upstream subscription (no NGR)
-  auto consumer1 = createMockConsumer();
-  auto handle1 = subscribeToTrack(subscriber1, kTestTrackName, consumer1, RequestID(1));
-  ASSERT_NE(handle1, nullptr);
-
-  // Second subscriber includes NEW_GROUP_REQUEST=8
-  auto consumer2 = createMockConsumer();
-  SubscribeRequest sub2;
-  sub2.fullTrackName = kTestTrackName;
-  sub2.requestID = RequestID(2);
-  sub2.locType = LocationType::LargestObject;
-  sub2.params.insertParam(
-      Parameter(folly::to_underlying(TrackRequestParamKey::NEW_GROUP_REQUEST), uint64_t(8))
-  );
-
-  // The relay must forward NGR=8 upstream via REQUEST_UPDATE
-  EXPECT_CALL(*upstreamHandle, requestUpdateCalled(_)).WillOnce([](const RequestUpdate& update) {
-    auto ngrValue = getFirstIntParam(update.params, TrackRequestParamKey::NEW_GROUP_REQUEST);
-    ASSERT_TRUE(ngrValue.has_value());
-    EXPECT_EQ(*ngrValue, 8);
-  });
-
-  std::shared_ptr<SubscriptionHandle> handle2{nullptr};
-  withSessionContext(subscriber2, [&]() {
-    auto task = relay_->subscribe(std::move(sub2), consumer2);
-    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-    ASSERT_TRUE(res.hasValue());
-    handle2 = *res;
-  });
-  exec_->drive();
-
-  // Register handle2 so removeSession(subscriber2) will unsubscribe it,
-  // allowing the forwarder to become empty and release the upstream handle.
-  getOrCreateMockState(subscriber2)->subscribeHandles.push_back(handle2);
-
-  removeSession(publisherSession);
-  removeSession(subscriber1);
-  removeSession(subscriber2);
-  exec_->drive();
-}
-
-// Relay test: A downstream subscriber sending REQUEST_UPDATE with
-// NEW_GROUP_REQUEST causes the relay to cascade the NGR upstream
-TEST_F(MoQRelayTest, RelayRequestUpdateNGRCascadedUpstream) {
-  auto publisherSession = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  doPublishNamespace(publisherSession, kTestNamespace);
-
-  // Upstream SubscribeOk advertises DYNAMIC_GROUPS = true
-  SubscribeOk upstreamOk;
-  upstreamOk.requestID = RequestID(100);
-  upstreamOk.trackAlias = TrackAlias(1);
-  upstreamOk.expires = std::chrono::milliseconds(0);
-  upstreamOk.groupOrder = GroupOrder::OldestFirst;
-  setPublisherDynamicGroups(upstreamOk, true);
-
-  auto upstreamHandle = std::make_shared<NiceMock<MockSubscriptionHandle>>(upstreamOk);
-  ON_CALL(*upstreamHandle, requestUpdateResult())
-      .WillByDefault(Return(folly::makeExpected<RequestError>(
-          RequestOk{RequestID(0), TrackRequestParameters(FrameType::REQUEST_OK), {}}
-      )));
-
-  EXPECT_CALL(*publisherSession, subscribe(_, _))
-      .WillOnce([&upstreamHandle](const auto& /*req*/, auto /*consumer*/) {
-        return folly::coro::makeTask<Publisher::SubscribeResult>(
-            folly::Expected<std::shared_ptr<SubscriptionHandle>, SubscribeError>(upstreamHandle)
-        );
-      });
-
-  // Subscribe downstream session - triggers upstream subscribe
-  auto consumer = createMockConsumer();
-  auto handle = subscribeToTrack(subscriberSession, kTestTrackName, consumer, RequestID(1));
-  ASSERT_NE(handle, nullptr);
-
-  auto* subscriber = dynamic_cast<MoQForwarder::Subscriber*>(handle.get());
-  ASSERT_NE(subscriber, nullptr);
-
-  // The relay must cascade NGR=9 upstream via REQUEST_UPDATE
-  EXPECT_CALL(*upstreamHandle, requestUpdateCalled(_)).WillOnce([](const RequestUpdate& update) {
-    auto ngrValue = getFirstIntParam(update.params, TrackRequestParamKey::NEW_GROUP_REQUEST);
-    ASSERT_TRUE(ngrValue.has_value());
-    EXPECT_EQ(*ngrValue, 9);
-  });
-
-  // Downstream subscriber sends REQUEST_UPDATE carrying NEW_GROUP_REQUEST=9
-  RequestUpdate update;
-  update.requestID = RequestID(2);
-  update.existingRequestID = RequestID(1);
-  update.params.insertParam(
-      Parameter(folly::to_underlying(TrackRequestParamKey::NEW_GROUP_REQUEST), uint64_t(9))
-  );
-  folly::coro::blockingWait(subscriber->requestUpdate(std::move(update)));
-  exec_->drive();
-
-  removeSession(publisherSession);
-  removeSession(subscriberSession);
-}
-
-// Unit test: Subscriber::onPublishOk postprocessing
-// Verifies that onPublishOk correctly updates:
-// 1. Subscriber range based on PublishOk fields
-// 2. shouldForward flag
-// 3. NEW_GROUP_REQUEST forwarding when it passes gating checks
-TEST_F(MoQRelayTest, SubscriberOnPublishOkPostprocessing) {
-  auto publisherSession = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  // Publish track and subscribe
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-  auto consumer = createMockConsumer();
-  auto handle = subscribeToTrack(subscriberSession, kTestTrackName, consumer, RequestID(1));
-  ASSERT_NE(handle, nullptr);
-
-  // Cast to access Subscriber internals
-  auto* subscriber = dynamic_cast<MoQForwarder::Subscriber*>(handle.get());
-  ASSERT_NE(subscriber, nullptr);
-
-  // Test 1: Range update from PublishOk
-  // Create a PublishOk with specific start/end locations
-  PublishOk pubOk1{
-      RequestID(1),                // requestID
-      true,                        // forward
-      0,                           // subscriberPriority
-      GroupOrder::OldestFirst,     // groupOrder
-      LocationType::AbsoluteStart, // locType
-      AbsoluteLocation{5, 0},      // start
-      uint64_t(15),                // endGroup
-      TrackRequestParameters(FrameType::PUBLISH_OK)
-  };
-
-  // Apply the postprocessing
-  subscriber->onPublishOk(pubOk1);
-
-  // Verify range was updated
-  EXPECT_EQ(subscriber->range.start.group, 5);
-
-  // Test 2: Forward flag update
-  subscriber->shouldForward = true;
-  PublishOk pubOk2{
-      RequestID(2),
-      false, // forward = false
-      0,
-      GroupOrder::OldestFirst,
-      LocationType::AbsoluteStart,
-      AbsoluteLocation{5, 0},
-      uint64_t(15),
-      TrackRequestParameters(FrameType::PUBLISH_OK)
-  };
-  subscriber->onPublishOk(pubOk2);
-  EXPECT_FALSE(subscriber->shouldForward) << "Forward flag should be updated to false";
-
-  subscriber->shouldForward = false;
-  PublishOk pubOk3{
-      RequestID(3),
-      true, // forward = true
-      0,
-      GroupOrder::OldestFirst,
-      LocationType::AbsoluteStart,
-      AbsoluteLocation{5, 0},
-      uint64_t(15),
-      TrackRequestParameters(FrameType::PUBLISH_OK)
-  };
-  subscriber->onPublishOk(pubOk3);
-  EXPECT_TRUE(subscriber->shouldForward) << "Forward flag should be updated to true";
-
-  // Test 3: NEW_GROUP_REQUEST forwarding via onPublishOk
-  // Enable dynamic groups and attach a callback to observe NGR fires
-  PublishRequest pub;
-  setPublisherDynamicGroups(pub, true);
-  subscriber->forwarder.setExtensions(pub.extensions);
-
-  auto cb = std::make_shared<TestNGRCallback>();
-  subscriber->forwarder.setCallback(cb);
-
-  // Build a PublishOk carrying NEW_GROUP_REQUEST=20
-  TrackRequestParameters ngrParams(FrameType::PUBLISH_OK);
-  ngrParams.insertParam(
-      Parameter(folly::to_underlying(TrackRequestParamKey::NEW_GROUP_REQUEST), uint64_t(20))
-  );
-  PublishOk pubOk4{
-      RequestID(4),
-      true,
-      0,
-      GroupOrder::OldestFirst,
-      LocationType::AbsoluteStart,
-      AbsoluteLocation{10, 0},
-      std::nullopt,
-      std::move(ngrParams)
-  };
-
-  // onPublishOk should fire the NGR callback for group 20
-  subscriber->onPublishOk(pubOk4);
-  ASSERT_EQ(cb->calls.size(), 1u);
-  EXPECT_EQ(cb->calls[0], 20u);
-  cb->calls.clear();
-
-  // outstanding=20: re-requesting group 20 is a no-op; group 21 fires
-  subscriber->forwarder.tryProcessNewGroupRequest(makeNGRParams(20));
-  EXPECT_TRUE(cb->calls.empty()) << "Group 20 already outstanding";
-  subscriber->forwarder.tryProcessNewGroupRequest(makeNGRParams(21));
-  ASSERT_EQ(cb->calls.size(), 1u);
-  EXPECT_EQ(cb->calls[0], 21u);
-
-  removeSession(publisherSession);
-  removeSession(subscriberSession);
-}
-
-// Test: Duplicate beginSubgroup with active consumers resets them and creates
-// new ones.
-// Sequence: publish, 2 subscribers, beginSubgroup, beginSubgroup again ->
-// first consumers get reset, both subscribers get new consumers.
-TEST_F(MoQRelayTest, DuplicateSubgroupReplacesActiveConsumers) {
-  auto publisherSession = createMockSession();
-  auto sub1 = createMockSession();
-  auto sub2 = createMockSession();
-
-  auto mockConsumer1 = createMockConsumer();
-  auto mockConsumer2 = createMockConsumer();
-
-  auto sg1v1 = createMockSubgroupConsumer();
-  auto sg2v1 = createMockSubgroupConsumer();
-  auto sg1v2 = createMockSubgroupConsumer();
-  auto sg2v2 = createMockSubgroupConsumer();
-
-  // First beginSubgroup gives v1 consumers; second call gives v2 consumers
-  EXPECT_CALL(*mockConsumer1, beginSubgroup(0, 0, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg1v1);
-      })
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg1v2);
-      });
-  EXPECT_CALL(*mockConsumer2, beginSubgroup(0, 0, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg2v1);
-      })
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg2v2);
-      });
-
-  // v1 consumers should be reset when duplicate arrives
-  EXPECT_CALL(*sg1v1, reset(ResetStreamErrorCode::CANCELLED)).Times(1);
-  EXPECT_CALL(*sg2v1, reset(ResetStreamErrorCode::CANCELLED)).Times(1);
-  // v2 consumers should not be reset during duplicate handling; they will be
-  // closed cleanly via endOfSubgroup before teardown
-  EXPECT_CALL(*sg1v2, reset(_)).Times(0);
-  EXPECT_CALL(*sg2v2, reset(_)).Times(0);
-
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-  subscribeToTrack(sub1, kTestTrackName, mockConsumer1, RequestID(1));
-  subscribeToTrack(sub2, kTestTrackName, mockConsumer2, RequestID(2));
-
-  auto sgForwarder1 = publishConsumer->beginSubgroup(0, 0, 0);
-  EXPECT_TRUE(sgForwarder1.hasValue());
-
-  // Duplicate beginSubgroup - should reset v1 consumers and return new
-  // forwarder
-  auto sgForwarder2 = publishConsumer->beginSubgroup(0, 0, 0);
-  EXPECT_TRUE(sgForwarder2.hasValue());
-  EXPECT_NE(sgForwarder1.value(), sgForwarder2.value());
-
-  // Close the new subgroup cleanly before teardown to avoid reset during
-  // cleanup
-  EXPECT_TRUE(sgForwarder2.value()->endOfSubgroup().hasValue());
-
-  removeSession(publisherSession);
-  removeSession(sub1);
-  removeSession(sub2);
-}
-
-// Test: Duplicate beginSubgroup after all subscribers have stop_sending'd
-// returns CANCELLED to propagate the signal back to the publisher.
-TEST_F(MoQRelayTest, DuplicateSubgroupCancelledWhenNoActiveConsumers) {
-  auto publisherSession = createMockSession();
-  auto subscriber = createMockSession();
-
-  auto mockConsumer = createMockConsumer();
-  auto mockSg = std::make_shared<NiceMock<MockSubgroupConsumer>>();
-
-  EXPECT_CALL(*mockConsumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(mockSg);
-      });
-
-  // Subscriber's object() returns CANCELLED to simulate stop_sending
-  EXPECT_CALL(*mockSg, object(_, _, _, _))
-      .WillOnce(
-          Return(folly::makeUnexpected(MoQPublishError(MoQPublishError::CANCELLED, "stop sending")))
-      );
-
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-  subscribeToTrack(subscriber, kTestTrackName, mockConsumer, RequestID(1));
-
-  auto sgRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(sgRes.hasValue());
-  auto sg = sgRes.value();
-
-  // Trigger stop_sending tombstone via CANCELLED error from object()
-  sg->object(0, nullptr, {}, false);
-
-  // Duplicate beginSubgroup - all consumers tombstoned, should return CANCELLED
-  auto dupRes = publishConsumer->beginSubgroup(0, 0, 0);
-  EXPECT_TRUE(dupRes.hasError());
-  EXPECT_EQ(dupRes.error().code, MoQPublishError::CANCELLED);
-
-  removeSession(publisherSession);
-  removeSession(subscriber);
-}
-
-// Test: Duplicate beginSubgroup with partial stop_sending - active subscriber
-// gets reset and new consumer; tombstoned subscriber is skipped.
-TEST_F(MoQRelayTest, DuplicateSubgroupSkipsTombstonedSubscriber) {
-  auto publisherSession = createMockSession();
-  auto subA = createMockSession();
-  auto subB = createMockSession();
-
-  auto consumerA = createMockConsumer();
-  auto consumerB = createMockConsumer();
-
-  auto sgAv1 = createMockSubgroupConsumer();
-  auto sgBv1 = createMockSubgroupConsumer();
-  auto sgAv2 = createMockSubgroupConsumer();
-
-  // First beginSubgroup: both A and B get consumers
-  EXPECT_CALL(*consumerA, beginSubgroup(0, 0, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sgAv1);
-      })
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sgAv2);
-      });
-  EXPECT_CALL(*consumerB, beginSubgroup(0, 0, _, _))
-      .WillOnce([&](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sgBv1);
-      });
-
-  // object() is forwarded to both A and B; sub A succeeds, sub B returns
-  // CANCELLED to simulate stop_sending
-  EXPECT_CALL(*sgAv1, object(_, _, _, _))
-      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-  EXPECT_CALL(*sgBv1, object(_, _, _, _))
-      .WillOnce(
-          Return(folly::makeUnexpected(MoQPublishError(MoQPublishError::CANCELLED, "stop sending")))
-      );
-
-  // On duplicate: sub A's v1 consumer gets reset; sub B is tombstoned (no
-  // reset)
-  EXPECT_CALL(*sgAv1, reset(ResetStreamErrorCode::CANCELLED)).Times(1);
-  EXPECT_CALL(*sgBv1, reset(_)).Times(0);
-  EXPECT_CALL(*sgAv2, reset(_)).Times(0);
-
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-  subscribeToTrack(subA, kTestTrackName, consumerA, RequestID(1));
-  subscribeToTrack(subB, kTestTrackName, consumerB, RequestID(2));
-
-  auto sgForwarder1 = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(sgForwarder1.hasValue());
-
-  // Trigger tombstone for sub B via CANCELLED from object()
-  sgForwarder1.value()->object(0, nullptr, {}, false);
-
-  // Duplicate beginSubgroup: sub A gets reset+new, sub B is skipped
-  // (tombstoned)
-  auto sgForwarder2 = publishConsumer->beginSubgroup(0, 0, 0);
-  EXPECT_TRUE(sgForwarder2.hasValue());
-  EXPECT_NE(sgForwarder1.value(), sgForwarder2.value());
-
-  // Close the new subgroup cleanly before teardown
-  EXPECT_TRUE(sgForwarder2.value()->endOfSubgroup().hasValue());
-
-  removeSession(publisherSession);
-  removeSession(subA);
-  removeSession(subB);
-}
-
-// ============================================================
-// Publish Replaces Subscribe Tests
-// ============================================================
-
-// Regression test: When a PUBLISH replaces a subscribe-path subscription, the
-// old forwarder's subscribers must receive publishDone, and the new
-// publish-path subscription must be fully functional (accepting data from the
-// new publisher).
-TEST_F(MoQRelayTest, PublishReplacesSubscribeDrainsOldAndServesNew) {
-  auto publisherSession = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  doPublishNamespace(publisherSession, kTestNamespace);
-
-  // Set up upstream subscribe that succeeds
-  SubscribeOk upstreamOk;
-  upstreamOk.requestID = RequestID(1);
-  upstreamOk.trackAlias = TrackAlias(1);
-  upstreamOk.expires = std::chrono::milliseconds(0);
-  upstreamOk.groupOrder = GroupOrder::OldestFirst;
-
-  EXPECT_CALL(*publisherSession, subscribe(_, _))
-      .WillOnce([upstreamOk](const auto& /*req*/, auto /*consumer*/) {
-        auto handle = std::make_shared<NiceMock<MockSubscriptionHandle>>(upstreamOk);
-        return folly::coro::makeTask<Publisher::SubscribeResult>(
-            folly::Expected<std::shared_ptr<SubscriptionHandle>, SubscribeError>(handle)
-        );
-      });
-
-  // Subscribe to the track (creates subscribe-path subscription)
-  auto oldConsumer = createMockConsumer();
-  bool publishDoneReceived = false;
-  EXPECT_CALL(*oldConsumer, publishDone(_)).WillOnce([&publishDoneReceived](const PublishDone&) {
-    publishDoneReceived = true;
-    return folly::makeExpected<MoQPublishError>(folly::unit);
-  });
-  auto handle = subscribeToTrack(
-      subscriberSession,
-      kTestTrackName,
-      oldConsumer,
-      RequestID(1),
-      /*addToState=*/false
-  );
-  ASSERT_NE(handle, nullptr);
-
-  // PUBLISH arrives for the same track — replaces subscribe-path subscription
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-  ASSERT_NE(publishConsumer, nullptr);
-
-  // Old subscriber must have been drained
-  EXPECT_TRUE(publishDoneReceived) << "Old subscribe-path subscriber should receive publishDone";
-
-  // New publish-path subscription should be functional: subscribe a new
-  // downstream consumer and verify it receives data from the publisher
-  auto newConsumer = createMockConsumer();
-  auto sg = createMockSubgroupConsumer();
-  EXPECT_CALL(*newConsumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([&sg](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-  EXPECT_CALL(*sg, endOfSubgroup()).WillOnce(Return(folly::unit));
-
-  subscribeToTrack(subscriberSession, kTestTrackName, newConsumer, RequestID(2));
-
-  auto sgRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(sgRes.hasValue());
-  EXPECT_TRUE(sgRes.value()->endOfSubgroup().hasValue());
-
-  removeSession(publisherSession);
-  removeSession(subscriberSession);
-}
-
-// Test: publishDone on a forwarder with a forward-only subscriber (null
-// trackConsumer) should not crash. Forward-only subscribers are created by
-// addSubscriber(session, forward) which sets trackConsumer to nullptr.
-// drainSubscriber unconditionally calls subscriber.trackConsumer->publishDone()
-// which crashes on null.
-TEST_F(MoQRelayTest, PublishDoneWithForwardOnlySubscriber) {
-  auto session = createMockSession();
-
-  auto forwarder = std::make_shared<MoQForwarder>(kTestTrackName, AbsoluteLocation{0, 0});
-
-  // Add a forward-only subscriber (null trackConsumer)
-  auto subscriber = forwarder->addSubscriber(session, /*forward=*/true);
-  ASSERT_NE(subscriber, nullptr);
-  EXPECT_EQ(subscriber->trackConsumer, nullptr);
-
-  // publishDone should not crash despite null trackConsumer
-  auto res = forwarder->publishDone(
-      PublishDone{RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher ended"}
-  );
-  EXPECT_TRUE(res.hasValue());
-
-  // The subscriber should have been removed
-  EXPECT_TRUE(forwarder->empty());
-}
-
-// Test: removeSubscriber with a PublishDone value on a forward-only subscriber
-// (null trackConsumer) should not crash.
-TEST_F(MoQRelayTest, RemoveForwardOnlySubscriberWithPublishDone) {
-  auto session = createMockSession();
-
-  auto forwarder = std::make_shared<MoQForwarder>(kTestTrackName, AbsoluteLocation{0, 0});
-
-  // Add a forward-only subscriber (null trackConsumer)
-  auto subscriber = forwarder->addSubscriber(session, /*forward=*/true);
-  ASSERT_NE(subscriber, nullptr);
-  EXPECT_EQ(subscriber->trackConsumer, nullptr);
-
-  // removeSubscriber with pubDone should not crash
-  forwarder->removeSubscriber(
-      session,
-      PublishDone{RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "session disconnect"},
-      "test"
-  );
-  EXPECT_TRUE(forwarder->empty());
-}
-
-// Test: forwardChanged must not crash when called after the publisher has
-// terminated (onPublishDone clears handle/upstream). We trigger forwardChanged
-// via Subscriber::requestUpdate changing forward from true→false (1→0
-// transition). The subscriber survives drain because it has an open subgroup.
-TEST_F(MoQRelayTest, ForwardChangedAfterPublisherTermination) {
-  auto publisherSession = createMockSession();
-  auto subSession = createMockSession();
-
-  doPublishNamespace(publisherSession, kTestNamespace);
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName);
-
-  // Subscriber with forward=true (default)
-  auto consumer = createMockConsumer();
-  auto handle = subscribeToTrack(subSession, kTestTrackName, consumer, RequestID(0));
-  ASSERT_NE(handle, nullptr);
-
-  // Begin a subgroup so the subscriber has open subgroups and survives drain
-  auto sg = createMockSubgroupConsumer();
-  EXPECT_CALL(*consumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([&sg](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-  auto subgroupRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroupRes.hasValue());
-
-  // Publisher terminates — onPublishDone clears handle/upstream.
-  // forwarder->publishDone sets draining and calls drainSubscriber, but the
-  // subscriber has an open subgroup so it stays (receivedPublishDone_=true).
-  EXPECT_CALL(*consumer, publishDone(_))
-      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-  publishConsumer->publishDone(
-      {RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher ended"}
-  );
-
-  // Subscriber sends requestUpdate changing forward from true→false.
-  // This calls removeForwardingSubscriber → forwardingSubscribers_ 1→0 →
-  // forwardChanged on relay callback. forwardChanged accesses
-  // subscription.upstream which was nulled by onPublishDone → crash.
-  RequestUpdate update;
-  update.requestID = RequestID(0);
-  update.forward = false;
-  auto task = handle->requestUpdate(std::move(update));
-  auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-  EXPECT_TRUE(res.hasValue());
-
-  // Clean up: reset the subgroup so subscriber can be fully removed
-  EXPECT_CALL(*sg, reset(_)).Times(1);
-  subgroupRes.value()->reset(ResetStreamErrorCode::CANCELLED);
-
-  removeSession(publisherSession);
-  removeSession(subSession);
-}
-
-// Test: fetch fallback to subscriptions_ after publisher termination must not
-// crash. When findPublishNamespaceSession returns null (no publishNamespace),
-// fetch falls back to subscriptions_. After onPublishDone, upstream is null
-// but the subscription entry remains if the forwarder has subscribers.
-TEST_F(MoQRelayTest, FetchAfterPublisherTermination) {
-  auto publisherSession = createMockSession();
-  auto subSession = createMockSession();
-  auto fetchSession = createMockSession();
-
-  // Publish WITHOUT publishNamespace so findPublishNamespaceSession returns null
-  // and fetch falls back to subscriptions_
-  auto publishConsumer = doPublish(publisherSession, kTestTrackName, /*addToState=*/false);
-
-  // Subscriber with open subgroup so subscription survives publisher drain
-  auto consumer = createMockConsumer();
-  auto sg = createMockSubgroupConsumer();
-  EXPECT_CALL(*consumer, beginSubgroup(0, 0, _, _))
-      .WillOnce([&sg](uint64_t, uint64_t, uint8_t, bool) {
-        return folly::makeExpected<MoQPublishError, std::shared_ptr<SubgroupConsumer>>(sg);
-      });
-  auto handle = subscribeToTrack(subSession, kTestTrackName, consumer, RequestID(0));
-  ASSERT_NE(handle, nullptr);
-
-  // Begin subgroup to keep subscriber alive
-  auto subgroupRes = publishConsumer->beginSubgroup(0, 0, 0);
-  ASSERT_TRUE(subgroupRes.hasValue());
-
-  // Publisher terminates — clears upstream but subscription stays
-  EXPECT_CALL(*consumer, publishDone(_))
-      .WillOnce(Return(folly::makeExpected<MoQPublishError>(folly::unit)));
-  publishConsumer->publishDone(
-      {RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher ended"}
-  );
-
-  // Fetch from a different session — falls back to subscriptions_, gets null
-  // upstream, then crashes at line 1011 dereferencing null upstreamSession
-  Fetch fetch(RequestID(0), kTestTrackName, AbsoluteLocation{0, 0}, AbsoluteLocation{1, 0});
-  auto fetchConsumer = std::make_shared<NiceMock<MockFetchConsumer>>();
-  withSessionContext(fetchSession, [&]() {
-    auto task = relay_->fetch(std::move(fetch), fetchConsumer);
-    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-    // Should return an error, not crash
-    EXPECT_FALSE(res.hasValue());
-    EXPECT_EQ(res.error().errorCode, FetchErrorCode::TRACK_NOT_EXIST);
-  });
-
-  // Clean up
-  EXPECT_CALL(*sg, reset(_)).Times(1);
-  subgroupRes.value()->reset(ResetStreamErrorCode::CANCELLED);
-  handle->unsubscribe();
-  removeSession(publisherSession);
-  removeSession(subSession);
-  removeSession(fetchSession);
-}
-
-// Bug: when a subscriber with forward=true joins a namespace whose track
-// forwarder is empty, the relay fires REQUEST_UPDATE twice — once explicitly
-// at the if(forwarder->empty()) site and once via forwardChanged() when
-// addSubscriber() increments numForwardingSubscribers from 0 to 1.
-TEST_F(MoQRelayTest, SubscribeNs_ForwardTrue_EmptyForwarder_SingleRequestUpdate) {
-  auto pubSession = createMockSession();
-  doPublishNamespace(pubSession, kTestNamespace);
-  auto mockHandle = makePublishHandle();
-  doPublishWithHandle(pubSession, kTestTrackName, mockHandle);
-
-  // Expect exactly one REQUEST_UPDATE(forward=true).
-  // Before the fix this fires twice.
-  EXPECT_CALL(*mockHandle, requestUpdateCalled(_)).Times(1).WillOnce([](const RequestUpdate& u) {
-    ASSERT_TRUE(u.forward.has_value());
-    EXPECT_TRUE(*u.forward);
-  });
-
-  auto subSession = createMockSession();
-  setupPublishSucceeds(subSession);
-  doSubscribeNamespaceWithForward(subSession, kTestNamespace, /*forward=*/true);
-
-  for (int i = 0; i < 5; i++) {
-    exec_->drive();
-  }
-
-  // Verify before cleanup — cleanup itself legitimately sends forward=false
-  // when the subscriber leaves and the forwarder drains.
-  ASSERT_TRUE(testing::Mock::VerifyAndClearExpectations(mockHandle.get()));
-
-  removeSession(subSession);
-  removeSession(pubSession);
-  for (int i = 0; i < 3; i++) {
-    exec_->drive();
-  }
-}
-
-// Bug: when a subscriber with forward=false joins a namespace whose track
-// forwarder is empty, the relay fires a spurious REQUEST_UPDATE(forward=false)
-// at the if(forwarder->empty()) site — even though the upstream is already at
-// forward=false (set by publish() which found no subscribers).
-TEST_F(MoQRelayTest, SubscribeNs_ForwardFalse_EmptyForwarder_NoRequestUpdate) {
-  auto pubSession = createMockSession();
-  doPublishNamespace(pubSession, kTestNamespace);
-  auto mockHandle = makePublishHandle();
-  doPublishWithHandle(pubSession, kTestTrackName, mockHandle);
-
-  // Expect no REQUEST_UPDATE at all.
-  // Before the fix this fires once with forward=false.
-  EXPECT_CALL(*mockHandle, requestUpdateCalled(_)).Times(0);
-
-  auto subSession = createMockSession();
-  setupPublishSucceeds(subSession);
-  doSubscribeNamespaceWithForward(subSession, kTestNamespace, /*forward=*/false);
-
-  for (int i = 0; i < 5; i++) {
-    exec_->drive();
-  }
-
-  ASSERT_TRUE(testing::Mock::VerifyAndClearExpectations(mockHandle.get()));
-
-  removeSession(subSession);
-  removeSession(pubSession);
-  for (int i = 0; i < 3; i++) {
-    exec_->drive();
-  }
-}
-
-// Bug: when a second subscriber with forward=true joins an existing PUBLISH-path
-// subscription (causing a 0→1 forwarding transition), the relay fires REQUEST_UPDATE
-// twice — once via forwardChanged() (which fires synchronously inside addSubscriber
-// via addForwardingSubscriber) and once via the explicit block at the end of the
-// subscribe() else-branch. Analogous to the subscribeNamespace bug fixed in this PR.
-TEST_F(MoQRelayTest, Subscribe_SecondForwardingSubscriber_SingleRequestUpdate) {
-  auto pubSession = createMockSession();
-  doPublishNamespace(pubSession, kTestNamespace);
-  auto mockHandle = makePublishHandle();
-  doPublishWithHandle(pubSession, kTestTrackName, mockHandle);
-
-  // S1 joins with forward=false — no REQUEST_UPDATE expected (no forwarding change).
-  auto s1 = createMockSession();
-  setupPublishSucceeds(s1);
-  {
-    SubscribeRequest sub;
-    sub.fullTrackName = kTestTrackName;
-    sub.requestID = RequestID(1);
-    sub.locType = LocationType::LargestObject;
-    sub.forward = false;
-    withSessionContext(s1, [&]() {
-      auto res = folly::coro::blockingWait(
-          relay_->subscribe(std::move(sub), createMockConsumer()),
-          exec_.get()
-      );
-      EXPECT_TRUE(res.hasValue());
-      if (res.hasValue()) {
-        getOrCreateMockState(s1)->subscribeHandles.push_back(*res);
-      }
-    });
-  }
-  for (int i = 0; i < 3; i++) {
-    exec_->drive();
-  }
-
-  // Now expect exactly ONE REQUEST_UPDATE(forward=true) when S2 joins.
-  // Before the fix this fires TWICE (forwardChanged + explicit block).
-  EXPECT_CALL(*mockHandle, requestUpdateCalled(_)).Times(1).WillOnce([](const RequestUpdate& u) {
-    ASSERT_TRUE(u.forward.has_value());
-    EXPECT_TRUE(*u.forward);
-  });
-
-  auto s2 = createMockSession();
-  setupPublishSucceeds(s2);
-  {
-    SubscribeRequest sub;
-    sub.fullTrackName = kTestTrackName;
-    sub.requestID = RequestID(2);
-    sub.locType = LocationType::LargestObject;
-    sub.forward = true;
-    withSessionContext(s2, [&]() {
-      auto res = folly::coro::blockingWait(
-          relay_->subscribe(std::move(sub), createMockConsumer()),
-          exec_.get()
-      );
-      EXPECT_TRUE(res.hasValue());
-      if (res.hasValue()) {
-        getOrCreateMockState(s2)->subscribeHandles.push_back(*res);
-      }
-    });
-  }
-  for (int i = 0; i < 5; i++) {
-    exec_->drive();
-  }
-
-  // Verify before cleanup (cleanup legitimately sends forward=false).
-  ASSERT_TRUE(testing::Mock::VerifyAndClearExpectations(mockHandle.get()));
-
-  removeSession(s2);
-  removeSession(s1);
-  removeSession(pubSession);
-  for (int i = 0; i < 3; i++) {
-    exec_->drive();
-  }
-}
-
-// ============================================================
-// Peer relay namespace loop prevention tests
-// ============================================================
-
-// Regression test: when a peer relay reconnects and subscribes to our
-// namespaces, we must not echo back namespaces that originally came FROM that
-// peer — doing so overwrites the real publisher in the namespace tree and
-// breaks data flow for downstream subscribers.
-//
-// Scenario (relay acts as sg-sin-2-1, upstream is jp-osa-1):
-//   1. jp-osa-1 (session1) announces namespace NS to our relay.
-//   2. session1 drops (old QUIC connection not yet reaped).
-//   3. jp-osa-1 reconnects (session2) and sends a peer SUBSCRIBE_NAMESPACE.
-//   4. Bug: the relay walks its tree, finds NS with sourceSession=session1,
-//      session1 != session2, and delivers NS back to session2 — echo loop.
-//   5. Fix: NS is tagged with sourcePeerID="jp-osa-1" so the peer ID check
-//      suppresses the delivery regardless of session identity.
-//
-// Production relays negotiate draft-16 (empty prefix allowed).  The delivery
-// path for draft-16 is synchronous: namespacePublishHandle->namespaceMsg().
-TEST_F(MoQRelayTest, PeerNamespaceNotEchoedBackOnReconnect) {
-  // Relay must have a relayID for peer detection to activate.
-  relay_ = std::make_shared<MoqxRelay>(config::CacheConfig{.maxCachedTracks = 0}, "sg-sin-2-1");
-  relay_->setAllowedNamespacePrefix(kAllowedPrefix);
-
-  // Step 1: session1 is the peer's old (unreaped) connection.  Inject NS as
-  // if it arrived from peer "jp-osa-1" via the bridge handle.
-  auto session1 = createMockSession();
-  auto bridgeHandle = makeNamespaceBridgeHandle(relay_, session1, "jp-osa-1");
-  bridgeHandle->namespaceMsg(kTestNamespace);
-  ASSERT_EQ(relay_->findPublishNamespaceSessions(kTestNamespace).size(), 1u);
-
-  // Step 2: jp-osa-1 reconnects as session2 and sends a peer SUBSCRIBE_NAMESPACE.
-  // Peer-to-peer sessions negotiate draft-16 (empty prefix is a 16+ feature).
-  auto session2 = createMockSession();
-  ON_CALL(*session2, getNegotiatedVersion())
-      .WillByDefault(Return(std::optional<uint64_t>(kVersionDraft16)));
-
-  // The relay delivers existing namespaces via namespacePublishHandle->namespaceMsg
-  // (draft-16 synchronous path).  Use a mock handle to detect any echo.
-  auto nsHandle = std::make_shared<NiceMock<MockNamespacePublishHandle>>();
-  bool echoedBack = false;
-  ON_CALL(*nsHandle, namespaceMsg(_)).WillByDefault([&echoedBack](const TrackNamespace&) {
-    echoedBack = true;
-  });
-
-  withSessionContext(session2, [&]() {
-    // makePeerSubNs("jp-osa-1") carries jp-osa-1's relay ID in the auth token.
-    auto task = relay_->subscribeNamespace(makePeerSubNs("jp-osa-1"), nsHandle);
-    folly::coro::blockingWait(std::move(task), exec_.get());
-  });
-
-  EXPECT_FALSE(echoedBack) << "Relay echoed a peer's own namespace back to it on reconnect";
-
-  removeSession(session1);
-  removeSession(session2);
-}
-
-// Complement: namespaces from LOCAL publishers (not from the peer) must still
-// be delivered when that peer subscribes.
-TEST_F(MoQRelayTest, LocalNamespaceDeliveredToPeerOnReconnect) {
-  relay_ = std::make_shared<MoqxRelay>(config::CacheConfig{.maxCachedTracks = 0}, "sg-sin-2-1");
-  relay_->setAllowedNamespacePrefix(kAllowedPrefix);
-
-  // Local publisher session announces kTestNamespace.
-  auto localPublisher = createMockSession();
-  doPublishNamespace(localPublisher, kTestNamespace);
-
-  // Peer "jp-osa-1" subscribes with a draft-16 session.
-  auto peerSession = createMockSession();
-  ON_CALL(*peerSession, getNegotiatedVersion())
-      .WillByDefault(Return(std::optional<uint64_t>(kVersionDraft16)));
-
-  auto nsHandle = std::make_shared<NiceMock<MockNamespacePublishHandle>>();
-  bool delivered = false;
-  ON_CALL(*nsHandle, namespaceMsg(_)).WillByDefault([&delivered](const TrackNamespace&) {
-    delivered = true;
-  });
-
-  withSessionContext(peerSession, [&]() {
-    auto task = relay_->subscribeNamespace(makePeerSubNs("jp-osa-1"), nsHandle);
-    folly::coro::blockingWait(std::move(task), exec_.get());
-  });
-  EXPECT_TRUE(delivered) << "Relay failed to deliver a local namespace to a peer subscriber";
-
-  removeSession(localPublisher);
-  removeSession(peerSession);
-}
-
-// A mock session that simulates a peer announcing peerNs when the relay
-// calls subscribeNamespace on it (the reciprocal leg).
-//
-// handleOut receives the bridge handle so the test can control its lifetime
-// independently of the session object — mirroring how SubNsStreamCallback
-// lives in a coroutine frame that is separate from (but tied to) the session.
-class PeerAnnounceSession : public NiceMock<MockMoQSession> {
-public:
-  explicit PeerAnnounceSession(
-      std::shared_ptr<MoQExecutor> exec,
-      TrackNamespace peerNs,
-      std::shared_ptr<Publisher::NamespacePublishHandle>& handleOut
-  )
-      : NiceMock<MockMoQSession>(std::move(exec)), peerNs_(std::move(peerNs)),
-        handleOut_(handleOut) {}
-
-  folly::coro::Task<Publisher::SubscribeNamespaceResult> subscribeNamespace(
-      SubscribeNamespace subNs,
-      std::shared_ptr<Publisher::NamespacePublishHandle> handle
-  ) override {
-    handleOut_ = handle;
-    if (handle) {
-      handle->namespaceMsg(peerNs_);
-    }
-    co_return std::make_shared<NiceMock<MockSubscribeNamespaceHandle>>(
-        SubscribeNamespaceOk{subNs.requestID}
-    );
-  }
-
-private:
-  TrackNamespace peerNs_;
-  std::shared_ptr<Publisher::NamespacePublishHandle>& handleOut_;
-};
-
-// Full production-path regression test: verifies that the peerID stored on
-// namespace nodes via the reciprocal bridge handle is the INCOMING peer's relay
-// ID (not our own relayID_).
-//
-// Bug: makeNamespaceBridgeHandle was passed relayID_ ("sg-sin-2-1") instead of
-// incomingPeerID ("jp-osa-1"), so nodes got sourcePeerID="sg-sin-2-1".  On
-// reconnect the check "sg-sin-2-1" != "jp-osa-1" is true and the namespace
-// was echoed back — the loop survived.
-//
-// Unlike PeerNamespaceNotEchoedBackOnReconnect (which injects the namespace
-// directly via makeNamespaceBridgeHandle), this test goes through the full
-// relay_->subscribeNamespace() production path so the bug in the call-site is
-// exercised.
-TEST_F(MoQRelayTest, PeerNamespaceNotEchoedBack_FullProductionPath) {
-  relay_ = std::make_shared<MoqxRelay>(config::CacheConfig{.maxCachedTracks = 0}, "sg-sin-2-1");
-  relay_->setAllowedNamespacePrefix(kAllowedPrefix);
-
-  // Step 1: jp-osa-1 connects as session1.  It will announce kTestNamespace
-  // to us via the reciprocal bridge handle when we subscribe to it.
-  // bridgeHandle1 is held here (not inside the session) to model the production
-  // lifetime: the handle lives in the coroutine frame, separate from the session.
-  std::shared_ptr<Publisher::NamespacePublishHandle> bridgeHandle1;
-  auto session1 = std::make_shared<PeerAnnounceSession>(exec_, kTestNamespace, bridgeHandle1);
-  ON_CALL(*session1, getNegotiatedVersion())
-      .WillByDefault(Return(std::optional<uint64_t>(kVersionDraft16)));
-  getOrCreateMockState(session1);
-
-  auto nsHandle1 = std::make_shared<NiceMock<MockNamespacePublishHandle>>();
-  withSessionContext(session1, [&]() {
-    auto task = relay_->subscribeNamespace(makePeerSubNs("jp-osa-1"), nsHandle1);
-    folly::coro::blockingWait(std::move(task), exec_.get());
-  });
-
-  // kTestNamespace should now be in the tree with sourcePeerID="jp-osa-1".
-  ASSERT_EQ(relay_->findPublishNamespaceSessions(kTestNamespace).size(), 1u);
-
-  // Step 2: jp-osa-1 reconnects as session2 and re-subscribes.
-  // kTestNamespace must NOT be echoed back.
-  auto session2 = createMockSession();
-  ON_CALL(*session2, getNegotiatedVersion())
-      .WillByDefault(Return(std::optional<uint64_t>(kVersionDraft16)));
-
-  auto nsHandle2 = std::make_shared<NiceMock<MockNamespacePublishHandle>>();
-  bool echoedBack = false;
-  ON_CALL(*nsHandle2, namespaceMsg(_)).WillByDefault([&echoedBack](const TrackNamespace&) {
-    echoedBack = true;
-  });
-
-  withSessionContext(session2, [&]() {
-    auto task = relay_->subscribeNamespace(makePeerSubNs("jp-osa-1"), nsHandle2);
-    folly::coro::blockingWait(std::move(task), exec_.get());
-  });
-
-  EXPECT_FALSE(echoedBack) << "Relay echoed peer namespace back on reconnect (production path)";
-
-  // Simulate session1's coroutine frame being destroyed before session cleanup.
-  bridgeHandle1.reset();
-  removeSession(session1);
-  removeSession(session2);
-}
-
-// ============================================================
-// Bridge handle cleanup tests
-// ============================================================
-
-// Regression test: when a bridge handle is destroyed (ungraceful session close)
-// without graceful namespaceDoneMsg calls, tree entries it created must be
-// cleaned up so stale sourceSession shared_ptrs don't keep dead session objects
-// alive and downstream subscribers receive NAMESPACE_DONE.
-TEST_F(MoQRelayTest, BridgeHandleDestructorCleansUpNamespaces) {
-  auto upstreamSession = createMockSession();
-
-  // Simulate the bridge path: create a handle and announce a namespace through
-  // it, as happens when the relay subscribes to an upstream/peer.
-  auto bridgeHandle = makeNamespaceBridgeHandle(relay_, upstreamSession, /*peerID=*/{});
-  bridgeHandle->namespaceMsg(kTestNamespace);
-  ASSERT_EQ(relay_->findPublishNamespaceSessions(kTestNamespace).size(), 1u);
-
-  // Drop the bridge handle without graceful namespaceDoneMsg — simulates
-  // ungraceful session close destroying SubNsStreamCallback.
-  bridgeHandle.reset();
-
-  // Tree entry must be gone.
-  EXPECT_EQ(relay_->findPublishNamespaceSessions(kTestNamespace).size(), 0u)
-      << "Stale namespace tree entry persists after bridge handle destruction";
-
-  removeSession(upstreamSession);
-}
-
-// Verify that when a new publisher takes over a namespace before the old
-// bridge handle is destroyed, the stale handle's destructor does NOT evict
-// the new publisher's entry (doPublishNamespaceDone guards on sourceSession).
-TEST_F(MoQRelayTest, BridgeHandleDestructorDoesNotEvictNewPublisher) {
-  auto session1 = createMockSession();
-  auto session2 = createMockSession();
-
-  // session1 announces NS via bridge handle.
-  auto bridgeHandle1 = makeNamespaceBridgeHandle(relay_, session1, /*peerID=*/{});
-  bridgeHandle1->namespaceMsg(kTestNamespace);
-  ASSERT_EQ(relay_->findPublishNamespaceSessions(kTestNamespace).size(), 1u);
-
-  // session2 takes over NS (conflict path evicts session1, sets sourceSession=session2).
-  auto bridgeHandle2 = makeNamespaceBridgeHandle(relay_, session2, /*peerID=*/{});
-  bridgeHandle2->namespaceMsg(kTestNamespace);
-  ASSERT_EQ(relay_->findPublishNamespaceSessions(kTestNamespace).size(), 1u);
-
-  // Now session1's handle is destroyed (ungraceful close detected late).
-  // It must NOT evict session2's entry.
-  bridgeHandle1.reset();
-
-  EXPECT_EQ(relay_->findPublishNamespaceSessions(kTestNamespace).size(), 1u)
-      << "Stale bridge handle destructor evicted the new publisher's entry";
-
-  bridgeHandle2.reset();
-  removeSession(session1);
-  removeSession(session2);
-}
-
-// Regression test: publisher reconnects while a subscribe coroutine is
-// suspended at co_await upstreamSession->subscribe().  The reconnect
-// (doPublishNamespace for publisher 2) erases the subscribe-path subscriptions_
-// entry (upstream == publisherSession1 == nodePtr->sourceSession), then
-// publish() for the same FTN creates a new entry whose promise is already
-// satisfied.  The subscribe scope guard then calls setException() on the
-// already-satisfied promise → folly::PromiseAlreadySatisfied →
-// ScopeGuardImplBase::terminate() → std::terminate (exit code 139).
-//
-// Without the fix: crashes.  With the fix: subscribe returns an error cleanly.
-TEST_F(MoQRelayTest, PublishReconnectDuringSubscribeScopeGuardCrash) {
-  auto publisherSession1 = createMockSession();
-  auto publisherSession2 = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  // Publisher 1 announces the namespace so subscribe() will find it as the
-  // upstream session and call publisherSession1->subscribe().
-  PublishNamespace pn;
-  pn.trackNamespace = kTestNamespace;
-  relay_->doPublishNamespace(pn, publisherSession1, nullptr);
-
-  // Configure publisher 1's subscribe() mock to simulate publisher reconnect
-  // inline (the relay calls this during co_await upstreamSession->subscribe()):
-  //   1. Publisher 2 re-announces the namespace — doPublishNamespace erases the
-  //      subscribe-path subscriptions_ entry because its upstream field equals
-  //      publisherSession1 == nodePtr->sourceSession.
-  //   2. Publisher 2 publishes the FTN — publish() emplaces a new subscriptions_
-  //      entry and immediately calls rsub.promise.setValue(folly::unit).
-  //   3. Returns an error, simulating the old session being cancelled.
-  // After the mock returns, the subscribe error path fires the scope guard which
-  // does subscriptions_.find(trackName), finds the new publish-path entry, and
-  // calls it->second.promise.setException() on an already-satisfied promise →
-  // PromiseAlreadySatisfied → terminate().
-  std::shared_ptr<TrackConsumer> pub2Consumer;
-  EXPECT_CALL(*publisherSession1, subscribe(_, _))
-      .WillOnce(
-          [this, publisherSession2, &pub2Consumer](SubscribeRequest, std::shared_ptr<TrackConsumer>)
-              -> folly::coro::Task<Publisher::SubscribeResult> {
-            // Step 1: publisher 2 takes over the namespace.
-            PublishNamespace pn2;
-            pn2.trackNamespace = kTestNamespace;
-            relay_->doPublishNamespace(pn2, publisherSession2, nullptr);
-
-            // Step 2: publisher 2 publishes the FTN, creating a new
-            // subscriptions_ entry with the promise already satisfied.
-            {
-              folly::RequestContextScopeGuard ctx;
-              folly::RequestContext::get()->setContextData(
-                  sessionRequestToken(),
-                  std::make_unique<MoQSession::MoQSessionRequestData>(publisherSession2)
-              );
-              PublishRequest pub;
-              pub.fullTrackName = kTestTrackName;
-              auto res = relay_->publish(std::move(pub), createMockSubscriptionHandle());
-              EXPECT_TRUE(res.hasValue()) << "publish in mock unexpectedly failed";
-              if (res.hasValue()) {
-                pub2Consumer = res->consumer;
-              }
-            }
-
-            // Step 3: return error — simulates the old upstream session being cancelled.
-            co_return folly::makeUnexpected(SubscribeError{
-                RequestID(0),
-                SubscribeErrorCode::INTERNAL_ERROR,
-                "upstream session cancelled"
-            });
-          }
-      );
-
-  withSessionContext(subscriberSession, [&]() {
-    SubscribeRequest sub;
-    sub.fullTrackName = kTestTrackName;
-    sub.requestID = RequestID(1);
-    sub.locType = LocationType::LargestObject;
-    auto result = folly::coro::blockingWait(
-        relay_->subscribe(std::move(sub), createMockConsumer()),
-        exec_.get()
-    );
-    // With the fix: subscribe returns an error without crashing.
-    // Without the fix: std::terminate is called before this assertion runs.
-    EXPECT_FALSE(result.hasValue()) << "subscribe should have failed (upstream cancelled)";
-  });
-
-  if (pub2Consumer) {
-    pub2Consumer->publishDone(
-        {RequestID(0), PublishDoneStatusCode::SESSION_CLOSED, 0, "test cleanup"}
-    );
-  }
-  relay_->doPublishNamespaceDone(kTestNamespace, publisherSession2);
-}
-
-// Same reconnect scenario but the upstream subscribe returns OK instead of an
-// error.  Without the fix, the crash moves from the scope guard to the success
-// path: after g.dismiss(), subscriptions_.find() returns the new publish-path
-// entry (promise already satisfied), and rsub.promise.setValue() throws
-// PromiseAlreadySatisfied, which propagates as an unhandled coroutine exception.
-// With the fix: subscribe returns SUBSCRIBE_ERROR "publisher reconnected".
-TEST_F(MoQRelayTest, PublishReconnectDuringSubscribeSuccessPathCrash) {
-  auto publisherSession1 = createMockSession();
-  auto publisherSession2 = createMockSession();
-  auto subscriberSession = createMockSession();
-
-  PublishNamespace pn;
-  pn.trackNamespace = kTestNamespace;
-  relay_->doPublishNamespace(pn, publisherSession1, nullptr);
-
-  std::shared_ptr<TrackConsumer> pub2Consumer;
-  EXPECT_CALL(*publisherSession1, subscribe(_, _))
-      .WillOnce(
-          [this,
-           publisherSession2,
-           &pub2Consumer](SubscribeRequest subReq, std::shared_ptr<TrackConsumer>)
-              -> folly::coro::Task<Publisher::SubscribeResult> {
-            PublishNamespace pn2;
-            pn2.trackNamespace = kTestNamespace;
-            relay_->doPublishNamespace(pn2, publisherSession2, nullptr);
-
-            {
-              folly::RequestContextScopeGuard ctx;
-              folly::RequestContext::get()->setContextData(
-                  sessionRequestToken(),
-                  std::make_unique<MoQSession::MoQSessionRequestData>(publisherSession2)
-              );
-              PublishRequest pub;
-              pub.fullTrackName = kTestTrackName;
-              auto res = relay_->publish(std::move(pub), createMockSubscriptionHandle());
-              EXPECT_TRUE(res.hasValue()) << "publish in mock unexpectedly failed";
-              if (res.hasValue()) {
-                pub2Consumer = res->consumer;
-              }
-            }
-
-            // Return success — the crash moves to rsub.promise.setValue() in the
-            // success path (after g.dismiss()), which finds the new publish-path
-            // entry whose promise is already satisfied.
-            SubscribeOk ok;
-            ok.requestID = subReq.requestID;
-            ok.trackAlias = TrackAlias(subReq.requestID.value);
-            ok.expires = std::chrono::milliseconds(0);
-            ok.groupOrder = GroupOrder::OldestFirst;
-            co_return std::make_shared<NiceMock<MockSubscriptionHandle>>(std::move(ok));
-          }
-      );
-
-  withSessionContext(subscriberSession, [&]() {
-    SubscribeRequest sub;
-    sub.fullTrackName = kTestTrackName;
-    sub.requestID = RequestID(1);
-    sub.locType = LocationType::LargestObject;
-    auto result = folly::coro::blockingWait(
-        relay_->subscribe(std::move(sub), createMockConsumer()),
-        exec_.get()
-    );
-    EXPECT_FALSE(result.hasValue()) << "subscribe should fail (publisher reconnected)";
-    if (result.hasError()) {
-      EXPECT_EQ(result.error().reasonPhrase, "publisher reconnected during subscribe");
-    }
-  });
-
-  if (pub2Consumer) {
-    pub2Consumer->publishDone(
-        {RequestID(0), PublishDoneStatusCode::SESSION_CLOSED, 0, "test cleanup"}
-    );
-  }
-  relay_->doPublishNamespaceDone(kTestNamespace, publisherSession2);
-}
-
-// Regression: after publishDone the namespace-tree node must be pruned when
-// the track was the only remaining content.  (Was a bug before unpublishTrack
-// gained a NodeMutationGuard; kept as a regression guard.)
-TEST_F(MoQRelayTest, PublishDonePrunesNamespaceTreeNode) {
-  auto publisher = createMockSession();
-
-  doPublishNamespace(publisher, kTestNamespace);
-  auto consumer = doPublish(publisher, kTestTrackName);
-
-  // Verify the publish is visible in the tree
-  auto state = relay_->findPublishState(kTestTrackName);
-  EXPECT_TRUE(state.nodeExists);
-  EXPECT_EQ(state.session, publisher);
-
-  // publishNamespaceDone — node stays alive because the track publish is still active
-  withSessionContext(publisher, [&]() {
-    getOrCreateMockState(publisher)->publishNamespaceHandles[0]->publishNamespaceDone();
-    getOrCreateMockState(publisher)->publishNamespaceHandles.clear();
-  });
-
-  state = relay_->findPublishState(kTestTrackName);
-  EXPECT_TRUE(state.nodeExists);
-  EXPECT_EQ(state.session, publisher);
-
-  // End the track publish — node should now be pruned
-  withSessionContext(publisher, [&]() {
-    consumer->publishDone(
-        {RequestID(0), PublishDoneStatusCode::SUBSCRIPTION_ENDED, 0, "publisher done"}
-    );
-  });
-
-  state = relay_->findPublishState(kTestTrackName);
-  EXPECT_EQ(state.session, nullptr);
-  EXPECT_FALSE(state.nodeExists) << "Node persists after publish ended; pruning did not run";
-
-  removeSession(publisher);
-}
-
-// Empty namespace: publishNamespace with an empty TrackNamespace must not crash.
-TEST_F(MoQRelayTest, EmptyNamespacePublishNamespaceDone) {
-  auto publisher = createMockSession();
-
-  TrackNamespace emptyNs{{}};
-  PublishNamespace ann;
-  ann.trackNamespace = emptyNs;
-  withSessionContext(publisher, [&]() {
-    auto task = relay_->publishNamespace(std::move(ann), nullptr);
-    auto res = folly::coro::blockingWait(std::move(task), exec_.get());
-    if (res.hasValue()) {
-      getOrCreateMockState(publisher)->publishNamespaceHandles.push_back(res.value());
-    }
-  });
-
-  removeSession(publisher);
-}
-
-} // namespace moxygen::test
